### 2021-05-13 #3278 @whedon
[PRE REVIEW]: covidregionaldata: Subnational data for COVID-19 epidemiology
**Submitting author:** @joseph-palmer (<a href="http://orcid.org/0000-0002-5593-9352">Joseph Palmer</a>)
**Repository:** <a href="https://github.com/epiforecasts/covidregionaldata" target ="_blank">https://github.com/epiforecasts/covidregionaldata</a>
**Version:** 0.9.1
