### 2021-05-13 379e41e
updated tests to at least what we had, now have extra 3 tests
