+++ DESCRIPTION
@@ -88,6 +88,7 @@ Imports:
     readxl,
     rlang,
     stringr,
+    testthat (>= 3.0.0),
     tibble,
     tidyr (>= 1.0.0),
     tidyselect,
@@ -103,7 +104,6 @@ Suggests:
     rvest,
     rworldmap,
     spelling,
-    testthat (>= 3.0.0),
     usethis
 VignetteBuilder:
     knitr
