### 2021-05-14 4f90ade
fixed bug casuing data_url lists to form dataframe not list
