+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   dataset-status: dataset-status.html
   quickstart: quickstart.html
   slowstart: slowstart.html
-last_built: 2021-05-14T01:27Z
+last_built: 2021-05-14T10:34Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
