### 2021-05-14 deb4113
moved testing methods into seperate file, set up to run on a clone of the class and made interactive
