+++ R/shared-methods.R
@@ -511,156 +511,59 @@ DataClass <- R6::R6Class(
       return(sum_df)
     },
 
+    #' @description Specific test for a specific class.
+    #' By default this method is empty as if any code is required it should be
+    #' defined in a child class specific `specific_tests` method.
+    #' @param self_copy R6class the object to test
+    specific_tests = function(self_copy) {
+
+    },
+
     #' @description Run tests on data class. Inherited by child classes so test
-    #' run through each class.
+    #' run through each class. Tests data can be downloaded (if requested),
+    #' cleaned, processed and returned. Tests run on a cloned copy of the target
+    #' class.
     #' @param download logical. To download the data (TRUE) or use a snapshot
     #' (FALSE). Defaults to FALSE.
-    #' @importFrom purrr walk map
-    #' @importFrom dplyr slice_tail
-    #' @importFrom testthat test_that expect_s3_class expect_true expect_error
-    test = function(download = FALSE) {
+    #' @param snapshot_path character_array the path to save the downloaded
+    #' @param copy logical To work on a copy of the class or the actual class
+    #' @param ... Additional parameters to pass to `specific_tests`
+    #' snapshot to. Defaults to TRUE.
+    #' examples
+    #' \dontrun{
+    #'  # set up a countries class, e.g. Italy
+    #'  a <- Italy$new(level = "1")
+    #'  # assuming you run this from the main covidregionaldata repo
+    #'  snapshot_path <- "tests/testthat/custom_data/Italy_level_1.rds"
+    #'  a$test(snapshot_path = snapshot_path)
+    #' }
+    test = function(download = FALSE, snapshot_path = "",
+                    copy = TRUE, ...) {
       data_name <- paste0(class(self)[1], " at level ", self$level)
-      raw_path <- paste0(
-        "custom_data/", class(self)[1],
-        "_level_", self$level, ".rds"
-      )
-      if (!file.exists(raw_path)) {
-        download <- TRUE
-      }
-      if (download) {
-        test_that(paste0(data_name, " downloads sucessfully"), {
-          self$download()
-          walk(self$data$raw, function(data) {
-            expect_s3_class(data, "data.frame")
-            expect_true(nrow(data) > 0)
-            expect_true(ncol(data) >= 2)
-          })
-        })
-        self$data$raw <- map(self$data$raw,
-          slice_tail,
-          n = 250
-        )
-        self$data$raw <- map(
-          self$data$raw,
-          ~ .[, 1:min(100, ncol(.))]
-        )
-        saveRDS(self$data$raw, raw_path)
+      if (copy) {
+        test_class <- self$clone()
       } else {
-        self$data$raw <- readRDS(raw_path)
+        test_class <- self
       }
-      test_that(paste0(data_name, " can be cleaned as expected"), {
-        self$clean()
-        expect_s3_class(self$data$clean, "data.frame")
-        expect_true(nrow(self$data$clean) > 0)
-        expect_true(ncol(self$data$clean) >= 2)
-        self$expect_clean_cols(self$data$clean)
-      })
-
-      test_that(
-        paste0(data_name, " can highlight available regions as expected"),
-        {
-          expect_error(self$available_regions(), NA)
-          expect_true(class(self$available_regions()) %in% "character")
-        }
+      test_download(
+        self = test_class,
+        download = download,
+        data_name = data_name,
+        snapshot_path = snapshot_path
       )
-
-      test_that(paste0(data_name, " can be processed as expected"), {
-        self$process()
-        expect_s3_class(self$data$processed, "data.frame")
-        expect_true(nrow(self$data$processed) > 0)
-        expect_true(ncol(self$data$processed) >= 2)
-        self$expect_processed_cols(self$data$processed, level = self$level)
[…]
