+++ R/testing_functions.R
@@ -1,0 +1,179 @@
+#' Expect clean columns
+#' @description Expect data has cleaned columns. Inherited by child
+#' classes so tests run through each class.
+#' @param data The data to check
+#' @export
+expect_clean_cols <- function(data, level) {
+  testthat::expect_s3_class(data[["date"]], "Date")
+  testthat::expect_type(data[["level_1_region"]], "character")
+  if (level == "2") {
+    testthat::expect_type(data[["level_2_region"]], "character")
+  }
+}
+
+#' Expect processed columns
+#' @description Expect data has processed columns. Inherited by child
+#' classes so tests run through each class.
+#' @param data The data to check
+#' @param level character_array the level of the data to check
+#' @param localised logical to check localised data or not, defaults to
+#' TRUE.
+#' @export
+expect_processed_cols <- function(data, level = "1", localised = TRUE) {
+  testthat::expect_s3_class(data[["date"]], "Date")
+  testthat::expect_type(data[["cases_new"]], "double")
+  testthat::expect_type(data[["cases_total"]], "double")
+  testthat::expect_type(data[["deaths_new"]], "double")
+  testthat::expect_type(data[["deaths_total"]], "double")
+  if (!localised) {
+    testthat::expect_type(data[["level_1_region"]], "character")
+    if (level == "2") {
+      testthat::expect_type(data[["level_2_region"]], "character")
+    }
+  }
+}
+
+#' Expect that columns contain data
+#' @description Expect data has cleaned columns. Inherited by child
+#' classes so tests run through each class.
+#' @param data_name character_array The name of the class and level to
+#' check
+#' @importFrom purrr map walk
+#' @importFrom dplyr filter
+#' @importFrom rlang !!
+#' @export
+expect_columns_contain_data <- function(self, data_name) {
+  cols_present <- function(col) {
+    if (length(self$source_data_cols[grep(
+      col, tolower(self$source_data_cols)
+    )]) > 0) {
+      return(paste0(col, c("_new", "_total")))
+    } else {
+      return(NULL)
+    }
+  }
+  cols <- c("cases", "deaths", "recovered", "test")
+  cols2check <- map(cols, cols_present)
+  cols2check <- unlist(cols2check)
+  walk(
+    cols2check,
+    ~ {
+      testthat::test_that(
+        paste0(data_name, "column '", .x, "' is not just composed of NA"),
+        {
+          testthat::expect_true(
+            nrow(self$data$processed %>% filter(!is.na(!!.x))) > 0
+          )
+        }
+      )
+    }
+  )
+}
+
+#' Expect download runs ok
+#' @description Test data can be downloaded if download = TRUE, or a requested
+#' snapshot file is not found, and store a snap shot at the snapshot_path. If an
+#' existing snapshot file is found then just load this data to use in future
+#' tests.
+#' @param self The R6Class country class object to perform checks
+#' @param download logical check to download or use a snapshot of the data
+#' @param data_name character_array The name of the class and level to
+#' check
+#' @param snapshot_path character_array the path to save the downloaded
+#' snapshot to.
+#' @importFrom purrr map walk
+#' @importFrom dplyr slice_tail
+#' @export
+test_download <- function(self, download, data_name, snapshot_path) {
+  if (!(nchar(snapshot_path))) {
+    snapshot_path <- paste0(
+      "custom_data/", class(self)[1],
+      "_level_", self$level, ".rds"
+    )
+  }
+  if (!file.exists(snapshot_path)) {
+    download <- TRUE
+  }
+  if (download) {
[…]
