+++ man/expect_processed_cols.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/testing_functions.R
+\name{expect_processed_cols}
+\alias{expect_processed_cols}
+\title{Expect processed columns}
+\usage{
+expect_processed_cols(data, level = "1", localised = TRUE)
+}
+\arguments{
+\item{data}{The data to check}
+
+\item{level}{character_array the level of the data to check}
+
+\item{localised}{logical to check localised data or not, defaults to
+TRUE.}
+}
+\description{
+Expect data has processed columns. Inherited by child
+classes so tests run through each class.
+}

[…]
