### 2021-05-14 3b59c98
updated news and moved testthat to suggested packages
