+++ man/DataClass.Rd
@@ -396,15 +396,7 @@ class.
 \item{\code{copy}}{logical To work on a copy of the class or the actual class}
 
 \item{\code{...}}{Additional parameters to pass to \code{specific_tests}
-snapshot to. Defaults to TRUE.
-examples
-\dontrun{
- # set up a countries class, e.g. Italy
- a <- Italy$new(level = "1")
- # assuming you run this from the main covidregionaldata repo
- snapshot_path <- "tests/testthat/custom_data/Italy_level_1.rds"
- a$test(snapshot_path = snapshot_path)
-}}
+snapshot to. Defaults to TRUE.}
 }
 \if{html}{\out{</div>}}
 }
