+++ README.Rmd
@@ -131,7 +131,12 @@ uk_nots %>%
   guides(col = guide_legend(title = "Region"))
 ```
 
- See `get_available_datasets()` for supported regions and subregional levels. For further examples see the [quick start vignette](https://github.com/epiforecasts/covidregionaldata/blob/master/vignettes/quickstart.Rmd). Additional subnational data are supported via the `JHU()` and `Google()` classes. Use the `available_regions()` method once these data have been downloaded and cleaned (see their examples) for subnational data they internally support.
+See `get_available_datasets()` for supported regions and subregional levels.
+For an updated view of dataset status check the
+[hosted page](https://epiforecasts.io/covidregionaldata/articles/dataset-status.html)
+or build the [dataset status vignette](vignettes)
+ 
+For further examples see the [quick start vignette](https://github.com/epiforecasts/covidregionaldata/blob/master/vignettes/quickstart.Rmd). Additional subnational data are supported via the `JHU()` and `Google()` classes. Use the `available_regions()` method once these data have been downloaded and cleaned (see their examples) for subnational data they internally support.
 
 ## Citation
 
