+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   dataset-status: dataset-status.html
   quickstart: quickstart.html
   slowstart: slowstart.html
-last_built: 2021-05-15T01:15Z
+last_built: 2021-05-16T01:13Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
