### 2021-05-16 #3278 @csoneson
👋🏻 @mponce0, @eguidotti - would you be interested in reviewing this submission for JOSS?

> covidregionaldata: Subnational data for COVID-19 epidemiology
