### 2021-05-16 #3278 @eguidotti
@csoneson Thanks but I think to be in conflict of interest with this work as I've implemented and published something very similar. I will not review according to your policies.

> The definition of a conflict of Interest in peer review is a circumstance that makes you “unable to make an impartial scientific judgment or evaluation.” (PNAS Conflict of Interest Policy). JOSS is concerned with avoiding any actual conflicts of interest, and being sufficiently transparent that we avoid the appearance of conflicts of interest as well.
