### 2021-05-17 #3278 @mponce0
> Thanks both! @mponce0 - we discussed it and we think that it is not a problem (assuming of course that you think you can make a "fair" assessment of the submission).

@csoneson Thanks! I will certainly try my best!
