+++ README.md
@@ -72,7 +72,7 @@ the temporary directory by default),
 
 ``` r
 start_using_memoise()
-#> Using a cache at: /var/folders/ss/817x_f3x2gnfr2gn2jrv_m3c0000gn/T//RtmpBAYQeN
+#> Using a cache at: /tmp/RtmpdTom4d
 ```
 
 To stop using `memoise` use,
@@ -97,7 +97,7 @@ the Google COVID-19 open data project), use:
 ``` r
 nots <- get_national_data()
 #> Downloading data from https://covid19.who.int/WHO-COVID-19-global-data.csv
-#> Rows: 118,263
+#> Rows: 118,675
 #> Columns: 8
 #> Delimiter: ","
 #> chr  [3]: Country_code, Country, WHO_region
@@ -109,22 +109,20 @@ nots <- get_national_data()
 #> Cleaning data
 #> Processing data
 nots
-#> # A tibble: 118,263 x 15
-#>    date       un_region who_region country        iso_code cases_new cases_total
-#>    <date>     <chr>     <chr>      <chr>          <chr>        <dbl>       <dbl>
-#>  1 2020-01-03 Asia      EMRO       Afghanistan    AF               0           0
-#>  2 2020-01-03 Europe    EURO       Albania        AL               0           0
-#>  3 2020-01-03 Africa    AFRO       Algeria        DZ               0           0
-#>  4 2020-01-03 Oceania   WPRO       American Samoa AS               0           0
-#>  5 2020-01-03 Europe    EURO       Andorra        AD               0           0
-#>  6 2020-01-03 Africa    AFRO       Angola         AO               0           0
-#>  7 2020-01-03 Americas  AMRO       Anguilla       AI               0           0
-#>  8 2020-01-03 Americas  AMRO       Antigua & Bar… AG               0           0
-#>  9 2020-01-03 Americas  AMRO       Argentina      AR               0           0
-#> 10 2020-01-03 Asia      EURO       Armenia        AM               0           0
-#> # … with 118,253 more rows, and 8 more variables: deaths_new <dbl>,
-#> #   deaths_total <dbl>, recovered_new <dbl>, recovered_total <dbl>,
-#> #   hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>, tested_total <dbl>
+#> # A tibble: 118,737 x 15
+#>    date       un_region who_region country           iso_code cases_new cases_total deaths_new deaths_total recovered_new recovered_total hosp_new hosp_total tested_new tested_total
+#>    <date>     <chr>     <chr>      <chr>             <chr>        <dbl>       <dbl>      <dbl>        <dbl>         <dbl>           <dbl>    <dbl>      <dbl>      <dbl>        <dbl>
+#>  1 2020-01-03 Asia      EMRO       Afghanistan       AF               0           0          0            0            NA              NA       NA         NA         NA           NA
+#>  2 2020-01-03 Europe    EURO       Albania           AL               0           0          0            0            NA              NA       NA         NA         NA           NA
+#>  3 2020-01-03 Africa    AFRO       Algeria           DZ               0           0          0            0            NA              NA       NA         NA         NA           NA
+#>  4 2020-01-03 Oceania   WPRO       American Samoa    AS               0           0          0            0            NA              NA       NA         NA         NA           NA
+#>  5 2020-01-03 Europe    EURO       Andorra           AD               0           0          0            0            NA              NA       NA         NA         NA           NA
+#>  6 2020-01-03 Africa    AFRO       Angola            AO               0           0          0            0            NA              NA       NA         NA         NA           NA
+#>  7 2020-01-03 Americas  AMRO       Anguilla          AI               0           0          0            0            NA              NA       NA         NA         NA           NA
[…]
