### 2021-05-17 19b4682
made more recomended changes
