### 2021-05-18 a6edc02
made param names more verbose and moved data_name into class field
