+++ R/test-DataClass.R
@@ -35,16 +35,14 @@ expect_processed_cols <- function(data, level = "1", localised = TRUE) {
 #' Test that cleaned columns contain data and are the correct types
 #' @description Expect data has cleaned columns. Inherited by child
 #' classes so tests run through each class.
-#' @param self The DataClass object (R6Class) object to perform checks on.
-#' @param data_name character_array The name of the class and level to
-#' check
+#' @param cntry_obj The DataClass object (R6Class) object to perform checks on.
 #' @importFrom purrr map walk
 #' @importFrom dplyr filter
 #' @importFrom rlang !!
-expect_columns_contain_data <- function(self, data_name) {
+expect_columns_contain_data <- function(cntry_obj) {
   cols_present <- function(col) {
-    if (length(self$source_data_cols[grep(
-      col, tolower(self$source_data_cols)
+    if (length(cntry_obj$source_data_cols[grep(
+      col, tolower(cntry_obj$source_data_cols)
     )]) > 0) {
       return(paste0(col, c("_new", "_total")))
     } else {
@@ -58,10 +56,12 @@ expect_columns_contain_data <- function(self, data_name) {
     cols2check,
     ~ {
       testthat::test_that(
-        paste0(data_name, "column '", .x, "' is not just composed of NA"),
+        paste0(
+          cntry_obj$data_name, "column '", .x, "' is not just composed of NA"
+        ),
         {
           testthat::expect_true(
-            nrow(self$data$processed %>% filter(!is.na(!!.x))) > 0
+            nrow(cntry_obj$data$processed %>% filter(!is.na(!!.x))) > 0
           )
         }
       )
@@ -74,63 +74,60 @@ expect_columns_contain_data <- function(self, data_name) {
 #' snapshot file is not found, and store a snap shot at the snapshot_path. If an
 #' existing snapshot file is found then just load this data to use in future
 #' tests.
-#' @param self The DataClass object (R6Class) object to perform checks on.
+#' @param cntry_obj The DataClass object (R6Class) object to perform checks on.
 #' @param download logical check to download or use a snapshot of the data
-#' @param data_name character_array The name of the class and level to
-#' check
 #' @param snapshot_path character_array the path to save the downloaded
 #' snapshot to.
 #' @importFrom purrr map walk
 #' @importFrom dplyr slice_tail
-test_download <- function(self, download, data_name, snapshot_path) {
-  if (!(nchar(snapshot_path))) {
-    snapshot_path <- paste0(
-      "custom_data/", class(self)[1],
-      "_level_", self$level, ".rds"
-    )
-  }
+test_download <- function(cntry_obj, download, snapshot_path) {
   if (!file.exists(snapshot_path)) {
     download <- TRUE
   }
   if (download) {
-    testthat::test_that(paste0(data_name, " downloads sucessfully"), {
-      self$download()
-      walk(self$data$raw, function(data) {
+    testthat::test_that(paste0(cntry_obj$data_name, " downloads sucessfully"), {
+      cntry_obj$download()
+      walk(cntry_obj$data$raw, function(data) {
         testthat::expect_s3_class(data, "data.frame")
         testthat::expect_true(nrow(data) > 0)
         testthat::expect_true(ncol(data) >= 2)
       })
     })
-    self$data$raw <- map(self$data$raw,
+    cntry_obj$data$raw <- map(cntry_obj$data$raw,
       slice_tail,
       n = 250
     )
-    self$data$raw <- map(
-      self$data$raw,
+    cntry_obj$data$raw <- map(
+      cntry_obj$data$raw,
       ~ .[, 1:min(100, ncol(.))]
     )
[…]
