+++ man/expect_processed_cols.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/testing_functions.R
+% Please edit documentation in R/test-DataClass.R
 \name{expect_processed_cols}
 \alias{expect_processed_cols}
 \title{Test that processed columns contain the correct data types}
