+++ man/test_return.Rd
@@ -1,16 +1,13 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/testing_functions.R
+% Please edit documentation in R/test-DataClass.R
 \name{test_return}
 \alias{test_return}
 \title{Test return method works correctly}
 \usage{
-test_return(self, data_name)
+test_return(cntry_obj)
 }
 \arguments{
-\item{self}{The DataClass object (R6Class) object to perform checks on.}
-
-\item{data_name}{character_array The name of the class and level to
-check}
+\item{cntry_obj}{The DataClass object (R6Class) object to perform checks on.}
 }
 \description{
 Test data can be returned correctly using the return method
