+++ man/expect_columns_contain_data.Rd
@@ -1,16 +1,13 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/testing_functions.R
+% Please edit documentation in R/test-DataClass.R
 \name{expect_columns_contain_data}
 \alias{expect_columns_contain_data}
 \title{Test that cleaned columns contain data and are the correct types}
 \usage{
-expect_columns_contain_data(self, data_name)
+expect_columns_contain_data(cntry_obj)
 }
 \arguments{
-\item{self}{The DataClass object (R6Class) object to perform checks on.}
-
-\item{data_name}{character_array The name of the class and level to
-check}
+\item{cntry_obj}{The DataClass object (R6Class) object to perform checks on.}
 }
 \description{
 Expect data has cleaned columns. Inherited by child

[…]
