### 2021-05-18 71308f1
added test_all setting to control what settings are tested.
