### 2021-05-18 97acf2e
updated testing for new functions
