+++ man/JRC.Rd
@@ -1,99 +1,39 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/JHU.R
-\name{JHU}
-\alias{JHU}
-\title{R6 Class containing specific attributes and methods for John Hopkins
-University data}
+% Please edit documentation in R/JRC.R
+\name{JRC}
+\alias{JRC}
+\title{R6 Class containing specific attributes and methods for European Commission's
+Joint Research Centre data}
 \source{
-\url{https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data}
+\url{https://github.com/ec-jrc/COVID-19}
 }
 \description{
-Attributes and methods for COVID-19 data used for the 2019
-Novel Coronavirus Visual Dashboard operated by the Johns Hopkins University
-Center for Systems Science and Engineering (JHU CSSE). Supported by ESRI
-Living Atlas Team and the Johns Hopkins University Applied Physics Lab
-(JHU APL)
-}
-\details{
-This dataset support both national and subnational data sources
-with national level data returned by default. Subnational data is supported
-for a subset of countries which can be found after cleaning using the
-\code{available_regions()} method, see the examples for more details. These data
-sets are sourced, cleaned, standardised by the JHU team so please see the
-source repository for further details. Note that unlike many other data sets
-this means methods applied to this source are not being applied to raw
-surveillance data but instead to already cleaned data. If using for
-analysis checking the JHU source for further details is advisable.
-
-If using this data please cite:
-"Dong E, Du H, Gardner L. An interactive web-based dashboard to track
-COVID-19 in real time.
-Lancet Inf Dis. 20(5):533-534. doi: 10.1016/S1473-3099(20)30120-1"
+Information for downloading, cleaning and processing COVID-19
+region data from the European Commission's Joint Research Centre
 }
 \examples{
-# nolint start
 \dontrun{
-# set up a data cache
-start_using_memoise()
-
-# get all countries data
-jhu <- JHU$new(level = "1", get = TRUE)
-jhu$return()
-
-# show available regions with data at the second level of interest
-jhu_level_2 <- JHU$new(level = "2")
-jhu_level_2$download()
-jhu_level_2$clean()
-jhu$available_regions()
-
-# get all region data for the uk
-jhu_level_2$filter("uk")
-jhu_level_2$process()
-jhu_level_2$return()
+national <- JRC$new(verbose = TRUE, steps = TRUE, get = TRUE)
+national$return()
 }
-# nolint end
 }
 \seealso{
-Aggregated data sources
-\code{\link{Covid19DataHub}},
-\code{\link{Google}}
-
 National data sources
 \code{\link{Covid19DataHub}},
 \code{\link{ECDC}},
 \code{\link{Google}},
-\code{\link{JRC}},
+\code{\link{JHU}},
 \code{\link{WHO}}
-
-Subnational data sources
-\code{\link{Belgium}},
-\code{\link{Brazil}},
-\code{\link{Canada}},
-\code{\link{Colombia}},
-\code{\link{Covid19DataHub}},
-\code{\link{Cuba}},
-\code{\link{France}},
-\code{\link{Germany}},
-\code{\link{Google}},
-\code{\link{India}},
-\code{\link{Italy}},
-\code{\link{Lithuania}},
-\code{\link{Mexico}},
-\code{\link{SouthAfrica}},
-\code{\link{UK}},
-\code{\link{USA}}
 }
-\concept{aggregations}
[…]
