+++ R/ECDC.R
@@ -122,7 +122,8 @@ ECDC <- R6::R6Class("ECDC",
     },
 
     #' @description Run additional tests on ECDC class. Tests ECDC has required
-    #' additional columns and that there is only one row per country.
+    #' additional columns and that there is only one row per country. Designed
+    #' to be ran from `test` and not ran directly.
     #' @param self_copy R6class the object to test
     #' @param ... Extra params passed to specific download functions
     #' @importFrom dplyr filter group_by tally
