+++ R/WHO.R
@@ -96,7 +96,7 @@ WHO <- R6::R6Class("WHO",
     },
 
     #' @description Run additional tests on WHO data. Tests that there is only
-    #' one row per country
+    #' one row per country. Designed to be ran from `test` and not ran directly.
     #' @param self_copy R6class the object to test
     #' @param ... Extra params passed to specific download functions
     #' @importFrom dplyr filter group_by tally
