+++ R/test-DataClass.R
@@ -3,6 +3,8 @@
 #' column is a character in the cleaned data (data$clean)
 #' @param data The clean data to check
 #' @param level character_array the level of the data to check
+#' @family tests
+#' @export
 expect_clean_cols <- function(data, level) {
   testthat::expect_s3_class(data[["date"]], "Date")
   level_region_str <- paste0("level_", level, "_region")
@@ -31,14 +33,16 @@ expect_processed_cols <- function(data, level = "1", localised = TRUE) {
 #' Test that cleaned columns contain data/
 #' @description Checks that cleaned columns cases, deaths, recovered and test
 #' (new and total) are not entirely composed of NAs.
-#' @param cntry_obj The DataClass object (R6Class) object to perform checks on.
+#' @param DataClass_obj The DataClass object (R6Class) to perform checks on.
 #' @importFrom purrr map walk
 #' @importFrom dplyr filter
 #' @importFrom rlang !!
-expect_columns_contain_data <- function(cntry_obj) {
+#' @family tests
+#' @export
+expect_columns_contain_data <- function(DataClass_obj) {
   cols_present <- function(col) {
-    if (length(cntry_obj$source_data_cols[grep(
-      col, tolower(cntry_obj$source_data_cols)
+    if (length(DataClass_obj$source_data_cols[grep(
+      col, tolower(DataClass_obj$source_data_cols)
     )]) > 0) {
       return(paste0(col, c("_new", "_total")))
     } else {
@@ -53,11 +57,12 @@ expect_columns_contain_data <- function(cntry_obj) {
     ~ {
       testthat::test_that(
         paste0(
-          cntry_obj$data_name, "column '", .x, "' is not just composed of NA"
+          DataClass_obj$data_name, "column '",
+          .x, "' is not just composed of NA"
         ),
         {
           testthat::expect_true(
-            nrow(cntry_obj$data$processed %>% filter(!is.na(!!.x))) > 0
+            nrow(DataClass_obj$data$processed %>% filter(!is.na(!!.x))) > 0
           )
         }
       )
@@ -66,39 +71,44 @@ expect_columns_contain_data <- function(cntry_obj) {
 }
 
 #' Test download method works correctly
-#' @description Test data can be downloaded if download = TRUE, or a requested
-#' snapshot file is not found, and store a snap shot at the snapshot_path. If an
-#' existing snapshot file is found then load this data to use in future tests.
-#' @param cntry_obj The DataClass object (R6Class) object to perform checks on.
-#' @param download logical check to download or use a snapshot of the data
+#' @description Test data can be downloaded if `download = TRUE`, or a requested
+#' snapshot file is not found, and store a snap shot in the `snapshot_dir`. If
+#' an existing snapshot file is found then load this data to use in future tests
+#' @param DataClass_obj The R6Class object to perform checks on.
+#' @param download Logical check to download or use a snapshot of the data
 #' @param snapshot_path character_array the path to save the downloaded
 #' snapshot to.
 #' @importFrom purrr map walk
 #' @importFrom dplyr slice_tail
-test_download <- function(cntry_obj, download, snapshot_path) {
+#' @family tests
+#' @export
+test_download <- function(DataClass_obj, download, snapshot_path) {
   if (!file.exists(snapshot_path)) {
     download <- TRUE
   }
   if (download) {
-    testthat::test_that(paste0(cntry_obj$data_name, " downloads sucessfully"), {
-      cntry_obj$download()
-      walk(cntry_obj$data$raw, function(data) {
[…]
