+++ vignettes/testing.Rmd
@@ -2,7 +2,7 @@
 title: "Testing this package"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Quick start}
+  %\VignetteIndexEntry{Testing}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
