### 2021-05-18 53a0746
updated testing vignette
