### 2021-05-18 e2f50ec
started netherlands class from template, WIP
