+++ man/Lithuania.Rd
@@ -145,6 +145,7 @@ Subnational data sources
 \code{\link{Italy}},
 \code{\link{JHU}},
 \code{\link{Mexico}},
+\code{\link{Netherlands}},
 \code{\link{SouthAfrica}},
 \code{\link{UK}},
 \code{\link{USA}}
