+++ man/Netherlands.Rd
@@ -1,18 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/India.R
-\name{India}
-\alias{India}
-\title{India Class for downloading, cleaning and processing notification data}
+% Please edit documentation in R/Netherlands.R
+\name{Netherlands}
+\alias{Netherlands}
+\title{Netherlands Class for downloading, cleaning and processing notification data}
 \source{
-\url{https://www.covid19india.org}
+\url{https://coronadashboard.government.nl/verantwoording#vaccination}
 }
 \description{
 Information for downloading, cleaning
-and processing COVID-19 region data for India.
+and processing COVID-19 region data for the Netherlands.
 }
 \examples{
 \dontrun{
-region <- India$new(verbose = TRUE, steps = TRUE, get = TRUE)
+region <- Netherlands$new(verbose = TRUE, steps = TRUE, get = TRUE)
 region$return()
 }
 }
@@ -27,11 +27,11 @@ Subnational data sources
 \code{\link{France}},
 \code{\link{Germany}},
 \code{\link{Google}},
+\code{\link{India}},
 \code{\link{Italy}},
 \code{\link{JHU}},
 \code{\link{Lithuania}},
 \code{\link{Mexico}},
-\code{\link{Netherlands}},
 \code{\link{SouthAfrica}},
 \code{\link{UK}},
 \code{\link{USA}}
@@ -39,7 +39,7 @@ Subnational data sources
 \concept{dataset}
 \concept{subnational}
 \section{Super class}{
-\code{\link[covidregionaldata:DataClass]{covidregionaldata::DataClass}} -> \code{India}
+\code{\link[covidregionaldata:DataClass]{covidregionaldata::DataClass}} -> \code{Netherlands}
 }
 \section{Public fields}{
 \if{html}{\out{<div class="r6-fields">}}
@@ -52,7 +52,8 @@ Subnational data sources
 
 \item{\code{supported_region_codes}}{A list of region codes in order of level.}
 
-\item{\code{common_data_urls}}{List of named links to raw data.}
+\item{\code{common_data_urls}}{List of named links to raw data. The first, and
+only entry, is be named main.}
 
 \item{\code{source_data_cols}}{existing columns within the raw data}
 }
@@ -61,10 +62,9 @@ Subnational data sources
 \section{Methods}{
 \subsection{Public methods}{
 \itemize{
-\item \href{#method-set_region_codes}{\code{India$set_region_codes()}}
-\item \href{#method-clean_common}{\code{India$clean_common()}}
-\item \href{#method-get_desired_status}{\code{India$get_desired_status()}}
-\item \href{#method-clone}{\code{India$clone()}}
+\item \href{#method-set_region_codes}{\code{Netherlands$set_region_codes()}}
+\item \href{#method-clean_common}{\code{Netherlands$clean_common()}}
+\item \href{#method-clone}{\code{Netherlands$clone()}}
 }
 }
 \if{html}{
@@ -88,7 +88,7 @@ Subnational data sources
 \subsection{Method \code{set_region_codes()}}{
 Set up a table of region codes for clean data
 \subsection{Usage}{
-\if{html}{\out{<div class="r">}}\preformatted{India$set_region_codes()}\if{html}{\out{</div>}}
+\if{html}{\out{<div class="r">}}\preformatted{Netherlands$set_region_codes()}\if{html}{\out{</div>}}
 }
 
 }
@@ -96,28 +96,11 @@ Set up a table of region codes for clean data
 \if{html}{\out{<a id="method-clean_common"></a>}}
[…]
