+++ man/Covid19DataHub.Rd
@@ -84,6 +84,7 @@ Subnational data sources
 \code{\link{JHU}},
 \code{\link{Lithuania}},
 \code{\link{Mexico}},
+\code{\link{Netherlands}},
 \code{\link{SouthAfrica}},
 \code{\link{UK}},
 \code{\link{USA}}
