+++ man/JHU.Rd
@@ -78,6 +78,7 @@ Subnational data sources
 \code{\link{Italy}},
 \code{\link{Lithuania}},
 \code{\link{Mexico}},
+\code{\link{Netherlands}},
 \code{\link{SouthAfrica}},
 \code{\link{UK}},
 \code{\link{USA}}
