+++ man/SouthAfrica.Rd
@@ -32,6 +32,7 @@ Subnational data sources
 \code{\link{JHU}},
 \code{\link{Lithuania}},
 \code{\link{Mexico}},
+\code{\link{Netherlands}},
 \code{\link{UK}},
 \code{\link{USA}}
 }
