+++ NAMESPACE
@@ -18,6 +18,7 @@ export(Italy)
 export(JHU)
 export(Lithuania)
 export(Mexico)
+export(Netherlands)
 export(SouthAfrica)
 export(UK)
 export(USA)
