### 2021-05-18 42f16e9
added test to check for workflow file for a class and add it if not found
