+++ inst/github_workflow_template.yaml
@@ -3,10 +3,10 @@ on:
     - cron: '36 12 * * *'
   workflow_dispatch:
 
-name: WHO
+name: _SOURCE_
 
 jobs:
-  WHO:
+  _SOURCE_:
     runs-on: macOS-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
@@ -42,7 +42,7 @@ jobs:
       - name: Test dataset
         run: |
           options("testDownload" = TRUE)
-          options("testSource" = "WHO")
+          options("testSource" = "_SOURCE_")
           devtools::load_all()
           testthat::test_file("tests/testthat/test-regional-datasets.R", reporter = c("summary", "fail"))
         shell: Rscript {0}
