+++ man/JRC.Rd
@@ -8,13 +8,28 @@ Joint Research Centre data}
 \url{https://github.com/ec-jrc/COVID-19}
 }
 \description{
-Information for downloading, cleaning and processing COVID-19
-region data from the European Commission's Joint Research Centre
+Class for downloading, cleaning and processing COVID-19
+region data from the European Commission's Joint Research Centre. Subnational
+data (admin level 1) on numbers of contagious and fatalities by COVID-19,
+collected directly from the National Authoritative sources (National
+monitoring websites, when available). For more details see
+https://github.com/ec-jrc/COVID-19
 }
 \examples{
 \dontrun{
-national <- JRC$new(verbose = TRUE, steps = TRUE, get = TRUE)
-national$return()
+# get country level data
+jrc_level_1 <- JRC$new(level = "1", verbose = TRUE, steps = TRUE, get = TRUE)
+jrc_level_1$return()
+
+# show available regions with data at the first level of interest (country)
+jrc_level_1$available_regions()
+
+# get region level data
+jrc_level_2 <- JRC$new(level = "2", verbose = TRUE, steps = TRUE, get = TRUE)
+jrc_level_2$return()
+
+# show available regions with data at the second level of interest (region)
+jrc_level_2$available_regions()
 }
 }
 \seealso{
@@ -76,7 +91,9 @@ National data sources
 \if{html}{\out{<a id="method-clean_common"></a>}}
 \if{latex}{\out{\hypertarget{method-clean_common}{}}}
 \subsection{Method \code{clean_common()}}{
-JRC specific data cleaning
+JRC specific data cleaning. The raw source data columns are
+converted to the correct type and renamed appropriately to match the
+standard for general processing.
 \subsection{Usage}{
 \if{html}{\out{<div class="r">}}\preformatted{JRC$clean_common()}\if{html}{\out{</div>}}
 }
@@ -86,8 +103,8 @@ JRC specific data cleaning
 \if{html}{\out{<a id="method-clean_level_1"></a>}}
 \if{latex}{\out{\hypertarget{method-clean_level_1}{}}}
 \subsection{Method \code{clean_level_1()}}{
-JRC specific country level data cleaning. Aggregates the
-data to the country (level 1) level.
+JRC specific country level data cleaning. Selects country
+level (level 1) columns from the data ready for further processing.
 \subsection{Usage}{
 \if{html}{\out{<div class="r">}}\preformatted{JRC$clean_level_1()}\if{html}{\out{</div>}}
 }
@@ -97,8 +114,9 @@ data to the country (level 1) level.
 \if{html}{\out{<a id="method-clean_level_2"></a>}}
 \if{latex}{\out{\hypertarget{method-clean_level_2}{}}}
 \subsection{Method \code{clean_level_2()}}{
-JRC specific region level data cleaning. Aggregates the
-data to the region (level 2) level.
+JRC specific region level data cleaning. Selects country
+(level 1) and region (level 2) columns from the data ready for further
+processing.
 \subsection{Usage}{
 \if{html}{\out{<div class="r">}}\preformatted{JRC$clean_level_2()}\if{html}{\out{</div>}}
 }
