+++ tests/testthat/test-github_workflow_present.R
@@ -23,8 +23,9 @@ test_github_workflow_found <- function(source) {
   )
 }
 
-skip_on_cran()
-purrr::walk(
-  get_available_datasets()$class,
-  test_github_workflow_found
-)
+if (identical(Sys.getenv("NOT_CRAN"), "true")) {
+  purrr::walk(
+    get_available_datasets()$class,
+    test_github_workflow_found
+  )
+}
