### 2021-05-19 9aab8ab
updated docs, added tmp dir for snapshots and updated uk nhs download
