+++ R/ECDC.R
@@ -123,7 +123,7 @@ ECDC <- R6::R6Class("ECDC",
 
     #' @description Run additional tests on ECDC class. Tests ECDC has required
     #' additional columns and that there is only one row per country. Designed
-    #' to be ran from `test` and not ran directly.
+    #' to be run from `test` and not run directly.
     #' @param self_copy R6class the object to test
     #' @param ... Extra params passed to specific download functions
     #' @importFrom dplyr filter group_by tally
