+++ man/test_processing.Rd
@@ -20,7 +20,7 @@ it is a data.frame, which is not empty, has at least 2 columns and calls
 \code{expect_processed_columns} to check each column types.
 }
 \seealso{
-Other tests: 
+Functions used for testing data is cleaned and processed correctly
 \code{\link{expect_clean_cols}()},
 \code{\link{expect_columns_contain_data}()},
 \code{\link{test_cleaning}()},
