### 2021-05-19 1c4101e
removed auto make from tests, added function to set up a new class which automatically creates a class from the template
