+++ tests/testthat/test-github_workflow_present.R
@@ -1,20 +1,10 @@
-make_github_workflow <- function(source, source_workflow_path) {
-  template_path <- "../../inst/github_workflow_template.yaml"
-  template <- readLines(template_path)
-  newfile <- gsub("_SOURCE_", source, template)
-  writeLines(newfile, source_workflow_path)
-  message(
-    paste("workflow auto created for", source, "at", source_workflow_path)
-  )
-}
-
 test_github_workflow_found <- function(source) {
   source_workflow_path <- paste0("../../.github/workflows/", source, ".yaml")
-  if (!file.exists(source_workflow_path)) {
-    make_github_workflow(source, source_workflow_path)
-  }
   test_that(
-    paste0(source, " has github workflow yaml file"),
+    paste0(
+      source, " has github workflow yaml file. ",
+      "You can create this by calling `make_github_workflow(`'", source, "')`."
+    ),
     {
       expect_true(
         file.exists(paste0("../../.github/workflows/", source, ".yaml"))
