+++ man/make_github_workflow.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{make_github_workflow}
+\alias{make_github_workflow}
+\title{Create github action for a given source}
+\usage{
+make_github_workflow(
+  source,
+  workflow_path = paste0(".github/workflows/", source, ".yaml")
+)
+}
+\arguments{
+\item{source}{character_array The name of the class to create the workflow
+for.}
+
+\item{workflow_path}{character_array The path to where the workflow file
+should be saved. Defaults to '.github/workflows/'}
+}
+\description{
+Makes a github workflow yaml file for a given source to be used
+as an action to check the data as a github action.
+}
