+++ man/expect_columns_contain_data.Rd
@@ -17,6 +17,7 @@ Checks that cleaned columns cases, deaths, recovered and test
 \seealso{
 Functions used for testing data is cleaned and processed correctly
 \code{\link{expect_clean_cols}()},
+\code{\link{expect_processed_cols}()},
 \code{\link{test_cleaning}()},
 \code{\link{test_download}()},
 \code{\link{test_processing}()},
