+++ man/test_cleaning.Rd
@@ -21,6 +21,7 @@ they are all characters.
 Functions used for testing data is cleaned and processed correctly
 \code{\link{expect_clean_cols}()},
 \code{\link{expect_columns_contain_data}()},
+\code{\link{expect_processed_cols}()},
 \code{\link{test_download}()},
 \code{\link{test_processing}()},
 \code{\link{test_return}()}
