+++ man/expect_processed_cols.Rd
@@ -18,3 +18,13 @@ TRUE.}
 Checks that processed data columns date, cases_new, cases_total,
 deaths_new, deaths_total and that region level have the correct types.
 }
+\seealso{
+Functions used for testing data is cleaned and processed correctly
+\code{\link{expect_clean_cols}()},
+\code{\link{expect_columns_contain_data}()},
+\code{\link{test_cleaning}()},
+\code{\link{test_download}()},
+\code{\link{test_processing}()},
+\code{\link{test_return}()}
+}
+\concept{tests}
