+++ man/test_download.Rd
@@ -24,6 +24,7 @@ an existing snapshot file is found then load this data to use in future tests
 Functions used for testing data is cleaned and processed correctly
 \code{\link{expect_clean_cols}()},
 \code{\link{expect_columns_contain_data}()},
+\code{\link{expect_processed_cols}()},
 \code{\link{test_cleaning}()},
 \code{\link{test_processing}()},
 \code{\link{test_return}()}
