+++ man/test_return.Rd
@@ -20,6 +20,7 @@ checked to ensure it contains data and is not just composed of NAs.
 Functions used for testing data is cleaned and processed correctly
 \code{\link{expect_clean_cols}()},
 \code{\link{expect_columns_contain_data}()},
+\code{\link{expect_processed_cols}()},
 \code{\link{test_cleaning}()},
 \code{\link{test_download}()},
 \code{\link{test_processing}()}
