+++ NAMESPACE
@@ -25,6 +25,7 @@ export(USA)
 export(WHO)
 export(expect_clean_cols)
 export(expect_columns_contain_data)
+export(expect_processed_cols)
 export(get_available_datasets)
 export(get_national_data)
 export(get_regional_data)
