### 2021-05-19 461d16e
Move testing into class (#343)
* moved dataset tests to a class method

* updated tests to at least what we had, now have extra 3 tests

* updated description with version increment and news with update overview

* moved testthat from suggested to imports

* forgot to use self when calling method

* forgot to use self when calling method

* updated uk nhs test to fix bug fixed in pr 346

* moved testing methods into seperate file, set up to run on a clone of the class and made interactive

* updated news and moved testthat to suggested packages

* removed who and ecdc from localised exclusion 
[…]
