+++ R/WHO.R
@@ -93,6 +93,21 @@ WHO <- R6::R6Class("WHO",
       } else {
         return(self$data$return)
       }
+    },
+
+    #' @description Run additional tests on WHO data. Tests that there is only
+    #' one row per country. Designed to be ran from `test` and not ran directly.
+    #' @param self_copy R6class the object to test
+    #' @param ... Extra params passed to specific download functions
+    #' @importFrom dplyr filter group_by tally
+    specific_tests = function(self_copy, ...) {
+      testthat::test_that("who data has expected format", {
+        all_countries <- self_copy$data$return %>%
+          filter(is.na(un_region)) %>%
+          group_by(country) %>%
+          tally()
+        testthat::expect_true(nrow(all_countries) == 0)
+      })
     }
   )
 )
