+++ man/test_processing.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/test-DataClass.R
+\name{test_processing}
+\alias{test_processing}
+\title{Test process method works correctly}
+\usage{
+test_processing(DataClass_obj, all = FALSE)
+}
+\arguments{
+\item{DataClass_obj}{The R6Class object to perform checks on.
+Must be a \code{DataClass} or \code{DataClass} child object.}
+
+\item{all}{Logical. Run tests with all settings (TRUE) or with those
+defined in the current class instance (FALSE). Defaults to FALSE.}
+}
+\description{
+Test data can be processed correctly using the process method.
+process is invoked to generate processed data which is then checked to ensure
+it is a data.frame, which is not empty, has at least 2 columns and calls
+\code{expect_processed_columns} to check each column types.
+}
+\seealso{
+Functions used for testing data is cleaned and processed correctly
+\code{\link{expect_clean_cols}()},
+\code{\link{expect_columns_contain_data}()},
+\code{\link{expect_processed_cols}()},
+\code{\link{test_cleaning}()},
+\code{\link{test_download}()},
+\code{\link{test_return}()}
+}
+\concept{tests}

[…]
