+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   quickstart: quickstart.html
   slowstart: slowstart.html
   testing: testing.html
-last_built: 2021-05-20T01:07Z
+last_built: 2021-05-20T17:05Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
