### 2021-05-21 bcc9da6
renamed region codes to be more informative
