+++ data-raw/jhu_codes.R
@@ -5,8 +5,8 @@ jhu_codes_link <- "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/mas
 # nolint end
 JHU_codes <- vroom(jhu_codes_link)
 JHU_codes <- JHU_codes %>%
-  select(Combined_Key, iso3) %>%
-  rename(
+  select(
+    level_1_region = Combined_Key,
     level_1_region_code = iso3
   )
 # update package region_codes
