### 2021-05-21 c30b5e9
added exports and updated NEWS
