+++ NAMESPACE
@@ -30,6 +30,8 @@ export(get_available_datasets)
 export(get_national_data)
 export(get_regional_data)
 export(initialise_dataclass)
+export(make_github_workflow)
+export(make_new_data_source)
 export(reset_cache)
 export(start_using_memoise)
 export(stop_using_memoise)
