### 2021-05-24 04c61c3
Updated README for JOSS comment
