+++ README.md
@@ -22,11 +22,16 @@ commits](https://img.shields.io/github/commits-since/epiforecasts/covidregionald
 [![DOI](https://zenodo.org/badge/271601189.svg)](https://zenodo.org/badge/latestdoi/271601189)
 [![status](https://joss.theoj.org/papers/dd6f7acdae3b7136a3ac373ce9a0655c/status.svg)](https://joss.theoj.org/papers/dd6f7acdae3b7136a3ac373ce9a0655c)
 
-An interface to subnational and national level COVID-19 data. For all
-countries supported, this includes a daily time-series of cases.
-Wherever available we also provide data on deaths, hospitalisations, and
-tests. National level data is also supported using a range of data
-sources as well as line list data and links to intervention data sets.
+`covidregionaldata` provides an interface to subnational and national
+level COVID-19 data sourced from both official sources, such as Public
+Health England (UK), and from other Covid-19 data collections, such as
+WHO and Google Open Data. For all countries supported, this includes a
+daily time-series of cases. Wherever available we also provide data on
+deaths, hospitalisations, and tests. National level data is also
+supported using a range of data sources as well as line list data and
+links to intervention data sets. Data is cleaned and processed from the
+raw sources on use in an open and transparent way, allowing users to
+scrutinise and extend our methods.
 
 ## Installation
 
@@ -72,7 +77,7 @@ the temporary directory by default),
 
 ``` r
 start_using_memoise()
-#> Using a cache at: /var/folders/ss/817x_f3x2gnfr2gn2jrv_m3c0000gn/T//RtmpBAYQeN
+#> Using a cache at: /tmp/RtmpN3Qhuk
 ```
 
 To stop using `memoise` use,
@@ -97,7 +102,7 @@ the Google COVID-19 open data project), use:
 ``` r
 nots <- get_national_data()
 #> Downloading data from https://covid19.who.int/WHO-COVID-19-global-data.csv
-#> Rows: 118,263
+#> Rows: 120,396
 #> Columns: 8
 #> Delimiter: ","
 #> chr  [3]: Country_code, Country, WHO_region
@@ -109,22 +114,20 @@ nots <- get_national_data()
 #> Cleaning data
 #> Processing data
 nots
-#> # A tibble: 118,263 x 15
-#>    date       un_region who_region country        iso_code cases_new cases_total
-#>    <date>     <chr>     <chr>      <chr>          <chr>        <dbl>       <dbl>
-#>  1 2020-01-03 Asia      EMRO       Afghanistan    AF               0           0
-#>  2 2020-01-03 Europe    EURO       Albania        AL               0           0
-#>  3 2020-01-03 Africa    AFRO       Algeria        DZ               0           0
-#>  4 2020-01-03 Oceania   WPRO       American Samoa AS               0           0
-#>  5 2020-01-03 Europe    EURO       Andorra        AD               0           0
-#>  6 2020-01-03 Africa    AFRO       Angola         AO               0           0
-#>  7 2020-01-03 Americas  AMRO       Anguilla       AI               0           0
-#>  8 2020-01-03 Americas  AMRO       Antigua & Bar… AG               0           0
-#>  9 2020-01-03 Americas  AMRO       Argentina      AR               0           0
-#> 10 2020-01-03 Asia      EURO       Armenia        AM               0           0
-#> # … with 118,253 more rows, and 8 more variables: deaths_new <dbl>,
-#> #   deaths_total <dbl>, recovered_new <dbl>, recovered_total <dbl>,
-#> #   hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>, tested_total <dbl>
+#> # A tibble: 120,396 x 15
[…]
