### 2021-05-24 #3290 @joseph-palmer
Thanks @mponce0 for the helpful review comments - really nice to get some critical eyes on the package. Some detailed responses below.

> LICENSE file: there are indeed two files named license, LICENSE and LICENSE.md, which are not consistent. Ie. one states the copyright and the other a proper MIT License.

The licence documentation follows the format of other packages, such as dplyr (https://github.com/tidyverse/dplyr). We are happy to update this is you have a link to better/newer practice that is CRAN compliant?

> Versioning, the version number of the package appears to be arbitrary, shouldn't be ver 1.0?

We follow the convention of MAJOR.MINOR.PATCH (https://semver.org/), with the current version (on branch master) being 0.9.2. As you say we will look to move to version 1.0 in the near future but are hoping for sizable user contributions to expand our coverage for subnational data sources in Africa and to refine our optional processing support.

>There is a good number of authors, including a "working group", which makes not clear the individual contributions. Perhaps a few sentences clarifying this can be added to the package repository or github page.

Contributions can be seen from the DESCRIPTION file and github commit logs (https://github.com/epiforecasts/covidregionaldata/pulse) which I believe shows clear evidence of code contributions. The working group refers to CMMID COVID-19 working group at LSHTM (https://cmmid.github.io/groups/ncov-group.html) and is referred to as members of the group made sizeable contributions to the planning and design of the package. @csoneson  what are the current JOSS guidelines for showing contributions to the package? It is unusual to have this in the package readme etc?

> It could be interesting to include some details about the data pipeline process, which appears to be done through github workflows.

The github workflows are used to check data integrity, such as that the raw data can be downloaded and that our inbuilt tests catch any changes, rather than to provide a processed database of data (unlike other packages that are similar). We felt that this was explained in the paper + docs but very happy to make changes to make this clearer?

> A statement of need: state what problems the software is designed to solve and who the target audience is

We have included a statement of need section in the paper which addresses what problem the software solves and who it is intended for. As requested we have extended our initial description in the README to give more detail (https://github.com/epiforecasts/covidregionaldata/pull/362).

> Although there are tests included, these do not appear to be run automatically when the package is installed, i.e. when using install.package() in R. Also R-CMD-check is showing a Failing status.

The package is passing checks both on CRAN and Github. For all packages (that we are aware of), tests run on package check rather than install. Users can check test status by either checking our data status page of the docs (https://epiforecasts.io/covidregionaldata/articles/dataset-status.html) or by running the tests directly (https://epiforecasts.io/covidregionaldata/articles/testing.html).

> Although the community guidelines are included in the documentation, the title of the section is a little but misleading, I'd call it "Community Contributions" or something alike…

The document outlines our code of conduct and standards and is adapted from the Contributor Covenant, version 2.0, available at https://www.contributor-covenant.org/version/2/0/code_of_conduct.html. We are hesitant to use ‘Community Contributions’ as this implies a page listing user contributions.

> Documentation could benefit from a few more examples, showing how to utilize more of the functions available in the package.

I agree, packages could always benefit from additional documentation but we have provided documentation for all of our functions, as well as vignettes outlining how to use the package. In an effort to expand our documentation we have currently open issues to develop our vignettes (https://github.com/epiforecasts/covidregionaldata/issues/316, https://github.com/epiforecasts/covidregionaldata/issues/314). If there is an area that you feel is critically lacking in documentation we can expand this but would need some more specific guidance.

> I'd also emphasize the link to the documentation site from the README file, as it is right now it can be easily missed.

The docs are currently linked in the README, and the cran site (https://cran.r-project.org/web/packages/covidregionaldata/index.html). We are happy to include additional links but a pointer indicating how best to make this visible would be helpful.

> One thing is not quite clear to me is how scalable is the solution of including new additional data sets, i.e. as stated in the manuscript "data sources which are often collected and maintained by individuals or small teams."

We have created guides to adding new data sources (https://github.com/epiforecasts/covidregionaldata/wiki) and data continues to be added to the package (https://github.com/epiforecasts/covidregionaldata/pull/358). We encourage our contributors to take ownership of the classes they add and monitor any changes which need addressing. Our automated data integrity checking allows us to identify datasets that are problematic without user reports. Of course if teams stop collating data then our processing will no longer be updated.

> Additionally, how robust are the data pipelines wrt to changes in or lost from the original sources. I.e. what would happen if a source link is broken or the data structure of the reported data changes, something that was quite frequent in some of the most popular data repository at the beginning of the pandemic.

As noted in the manuscript and in the package documentation we test all processing code on all datasets each night using GitHub actions. Changes in underlying data structure would cause these tests to fail (in core variables only). This would then be broadcast to users via the data status vignette and our rendered documentation. At this point either a data owner or a user would submit a patch to be added to the dev build of the package. Our docs would then show the issue was fixed in the dev version and not CRAN until we pushed that version to CRAN. Users can also run these tests themselves when interacting with the package.

Once again, big thanks for taking the time with this!
