### 2021-05-25 #3290 @federicomarini
Hi there @joseph-palmer, I'm going to start from some comments on the manuscript, while I continue with the code&package review.

Paper's well written and clearly describing the aims of `covidregionaldata` in the context of the need for fine-grained resolution data for localized health policy making for the COVID-19 pandemic.

An adequate survey of the current alternatives and state of the art is included, and the gap filled by this package is nicely defined.

A couple of minor comments:

- Summary:
  - maybe use a newer reference to the R core team?
  - l 16-17, repetition of extensible/extend
  - should links be kept also in their full form to be visible when printed? Might depend on the guidelines of JOSS, so I'm pinging @csoneson briefly

- Statement of need:
  - l.38: "changes to a dependency package" is not so clearly defined, do you mean something on the line of "changes in the subsequent processing steps" (for packages that use these sources)?
  - l.53: "test and report... daily" -> You could specify this happens via GHA and their output is recorded on the project webpage
  - while reading l.42-47: as a general idea: is there some way to track/record the origin or snapshot of the data for the files that are sourced? Something like an attribute could do it, but maybe some similar mechanism is already in place

- References:
  - Are the versions for `sars2pack` and `covdata` needed to ensure that this is the status quo at the moment of writing the manuscript? If dois are available, that could be even better

- Footnote: is it correct to have a `$` after `Palmer` as in `Palmer$ et al`? Maybe it is a feature of the build system
