### 2021-05-25 #3290 @csoneson
Thanks @federicomarini! On this:

> should links be kept also in their full form to be visible when printed? Might depend on the guidelines of JOSS, so I'm pinging @csoneson briefly

It's quite common in JOSS papers, so I'd say that's ok.
