### 2021-05-26 #3290 @federicomarini
Here's the rest of my review - I second most points raised by @mponce0, thanks for your thoughtful review as well!

The main point I'd raise is that a package user (let's say "the average one", not only a developer) could benefit a big lot from these fully worked out examples (also raised as a point in the checklist-driven review). Maybe I am _a little spoiled_ by how comprehensive some vignettes are in the Bioconductor project 😉

**Functionality**

- Installation: The installation works smoothly, as documented!
- Functionality: The package does deliver the functionality stated in the manuscript and is showcased in the vignettes

- Performance: No particular performance claims are made, but it seems that the runtime is reasonable - optimizations with `memoise` are in place to avoid repetitions of operations

**Documentation**

- A statement of need: These are elaborated in the vignettes. Maybe add a (more explicit) statement regarding what the target audience would be?
- Installation instructions: The dependencies are all installed and correctly handled - maybe you can point out that to run the vignette examples you might need to use `dependencies = TRUE` in the installation command?
- Example usage: The vignettes are present, and a very welcome piece of documentation. It might be overkill, but maybe some "fully worked out use cases" in one of the studies where the package has been used can add even more value? I realize there might be also some runtime/check-time constraints, so it is up to the authors to explore this point.
- Functionality documentation: Well documented!
- Automated tests: The test suite is well structured, and covers a wide portion of the code to retrieve and process the data. Maybe some additional tests might be desired to increase the overall coverage? Again, it might be a tradeoff between test runtimes and package coverage...
- Community guidelines: A code of conduct is present, and the BugReports field of the DESCRIPTION correctly points to the repo issues.

**Miscellaneous items**

- the help in some classes seems redundant, see ?Belgium (multiple same urls)
- slowstart has a function that returns an error - probably `country`?
- get_regional_data("Belgium", totals = TRUE) has an extra row - maybe worth run these chunks, or is it because of CRAN's policy?
- maybe have the slowstart vignette `eval`uated? By having these chunks automatically generated, it would be robust to changes in the underlying data
