### 2021-05-27 6a653c0
Initial commit adding Switzerland to covidregionaldata
