+++ man/Switzerland.Rd
@@ -1,18 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/India.R
-\name{India}
-\alias{India}
-\title{India Class for downloading, cleaning and processing notification data}
+% Please edit documentation in R/Switzerland.R
+\name{Switzerland}
+\alias{Switzerland}
+\title{Switzerland Class for downloading, cleaning and processing notification data}
 \source{
-\url{https://www.covid19india.org}
+\url{https://github.com/openZH/covid_19/}
 }
 \description{
 Information for downloading, cleaning
-and processing COVID-19 region data for India.
+and processing COVID-19 region data for Switzerland
 }
 \examples{
 \dontrun{
-region <- India$new(verbose = TRUE, steps = TRUE, get = TRUE)
+region <- Switzerland$new(verbose = TRUE, steps = TRUE, get = TRUE)
 region$return()
 }
 }
@@ -27,20 +27,20 @@ Subnational data sources
 \code{\link{France}},
 \code{\link{Germany}},
 \code{\link{Google}},
+\code{\link{India}},
 \code{\link{Italy}},
 \code{\link{JHU}},
 \code{\link{Lithuania}},
 \code{\link{Mexico}},
 \code{\link{Netherlands}},
 \code{\link{SouthAfrica}},
-\code{\link{Switzerland}},
 \code{\link{UK}},
 \code{\link{USA}}
 }
 \concept{dataset}
 \concept{subnational}
 \section{Super class}{
-\code{\link[covidregionaldata:DataClass]{covidregionaldata::DataClass}} -> \code{India}
+\code{\link[covidregionaldata:DataClass]{covidregionaldata::DataClass}} -> \code{Switzerland}
 }
 \section{Public fields}{
 \if{html}{\out{<div class="r6-fields">}}
@@ -62,10 +62,9 @@ Subnational data sources
 \section{Methods}{
 \subsection{Public methods}{
 \itemize{
-\item \href{#method-set_region_codes}{\code{India$set_region_codes()}}
-\item \href{#method-clean_common}{\code{India$clean_common()}}
-\item \href{#method-get_desired_status}{\code{India$get_desired_status()}}
-\item \href{#method-clone}{\code{India$clone()}}
+\item \href{#method-set_region_codes}{\code{Switzerland$set_region_codes()}}
+\item \href{#method-clean_common}{\code{Switzerland$clean_common()}}
+\item \href{#method-clone}{\code{Switzerland$clone()}}
 }
 }
 \if{html}{
@@ -90,7 +89,7 @@ Subnational data sources
 \subsection{Method \code{set_region_codes()}}{
 Set up a table of region codes for clean data
 \subsection{Usage}{
-\if{html}{\out{<div class="r">}}\preformatted{India$set_region_codes()}\if{html}{\out{</div>}}
+\if{html}{\out{<div class="r">}}\preformatted{Switzerland$set_region_codes()}\if{html}{\out{</div>}}
 }
 
 }
@@ -98,28 +97,11 @@ Set up a table of region codes for clean data
 \if{html}{\out{<a id="method-clean_common"></a>}}
 \if{latex}{\out{\hypertarget{method-clean_common}{}}}
 \subsection{Method \code{clean_common()}}{
-India state level data cleaning
+Switzerland specific state level data cleaning
 \subsection{Usage}{
-\if{html}{\out{<div class="r">}}\preformatted{India$clean_common()}\if{html}{\out{</div>}}
+\if{html}{\out{<div class="r">}}\preformatted{Switzerland$clean_common()}\if{html}{\out{</div>}}
 }
[…]
