+++ man/Cuba.Rd
@@ -33,6 +33,7 @@ Subnational data sources
 \code{\link{Mexico}},
 \code{\link{Netherlands}},
 \code{\link{SouthAfrica}},
+\code{\link{Switzerland}},
 \code{\link{UK}},
 \code{\link{USA}}
 }
