### 2021-05-31 0f8c275
Create standalone covr with ubuntu
