+++ .github/workflows/check-covr.yaml
@@ -61,7 +61,7 @@ jobs:
 
       - name: Install dependencies
         run: |
-            install.packages(c("remotes"))
+            install.packages(c("remotes","sessioninfo"))
             remotes::install_deps(dependencies = TRUE)
             remotes::install_cran("covr")
         shell: Rscript {0}
