### 2021-06-01 9e79dc3
made other recomended changes
