### 2021-06-01 257bbd8
Documentation, put Liechtenstein back in, added to News
