### 2021-06-02 #3290 @csoneson
Thanks @federicomarini and @mponce0 for your quick reactions!

@joseph-palmer - over to you now, please let us know if you have any questions!

And sorry for missing the ping above re: contributions - in general, JOSS leaves it up to the authors to decide on the [authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship) of the JOSS submission (assuming that this has been discussed and agreed upon among contributors). On the software side itself I think the first step is indeed to include the contributors in the DESCRIPTION file together with the type of contribution, as is currently done. Any additional clarifications, acknowledgements etc would, in my opinion, fit well in the README.
@joseph-palmer - I had just one question regarding the working group: you say above that it was included since "members of the group made sizeable contributions to the planning and design of the package". I also note that several of the named authors are indeed members of this working group. Could you clarify whether there were additional contributions made by the working group (beyond the named individuals)? I'm checking in with the larger editorial team as well, I don't think we have a lot of previous JOSS papers listing working groups or consortia among the authors :)
