### 2021-06-02 b0a896e
Apply suggestions on canton names from code review
Switching to more familiar/common forms of names for two Swiss cantons.

Co-authored-by: Stefan Oderbolz <oderbolz@gmail.com>
