+++ R/Switzerland.R
@@ -79,9 +79,9 @@ Switzerland <- R6::R6Class("Switzerland",
           "Appenzell Innerrhoden", "Basel-Landschaft",
           "Basel-Stadt", "Bern",
           "Fribourg", "Gen\u00e8ve",
-          "Glarus", "Grischun",
+          "Glarus", "Grisons",
           "Jura", "Luzern", "Neuch\u00e2tel", "Nidwalden",
-          "Obwalden", "Sankt Gallen", "Schaffhausen",
+          "Obwalden", "St. Gallen", "Schaffhausen",
           "Schwyz", "Solothurn", "Thurgau", "Ticino",
           "Uri", "Valais", "Vaud", "Zug",
           "Z\u00fcrich", "Liechtenstein"
