### 2021-06-04 b460ee7
Check multiple dates for working NHS data
