+++ R/UK.R
@@ -355,6 +355,8 @@ UK <- R6::R6Class("UK",
     #' @source \url{https://coronavirus.data.gov.uk/details/download}
     # nolint end
     #' @importFrom lubridate year month
+    #' @importFrom httr GET status_code
+    #' @importFrom purrr map_chr
     #' @importFrom readxl cell_limits
     #' @importFrom dplyr %>%
     download_nhs_regions = function() {
@@ -376,26 +378,42 @@ UK <- R6::R6Class("UK",
       )
 
       # 1. Data from 7 April 2021 to now:
-      nhs_recent_url <- paste0(
-        self$data_urls[["nhs_recent_url"]],
-        "/wp-content/uploads/sites/2/",
-        year(self$release_date), "/",
-        ifelse(month(self$release_date) < 10,
-          paste0(0, month(self$release_date)),
-          month(self$release_date)
-        ),
-        "/COVID-19-daily-admissions-and-beds-",
-        gsub("-", "", as.character(self$release_date)),
-        ".xlsx"
-      )
-
+      # Data not always daily; set up to try different dates
+      check_latest_date <- function(try_date) {
+        nhs_recent_url <- paste0(
+          self$data_urls[["nhs_recent_url"]],
+          "/wp-content/uploads/sites/2/",
+          year(try_date), "/",
+          ifelse(month(try_date) < 10,
+                 paste0(0, month(try_date)),
+                 month(try_date)
+          ),
+          "/COVID-19-daily-admissions-and-beds-",
+          gsub("-", "", as.character(try_date)),
+          ".xlsx")
+      }
+      # Get urls for last 7 days
+      try_date_seq <- seq.Date(self$release_date, 
+                           by = -1, length.out = 7)
+      try_urls <- map_chr(try_date_seq, 
+                          ~ check_latest_date(.x))
+      names(try_urls) <- try_date_seq
+      # Check for working urls
+      url_status <- map_chr(try_urls, 
+                        ~ GET(.x) %>%
+                          status_code())
+      # Keep latest working url
+      url_status <- url_status[(url_status == 200)]
+      names(url_status) <- as.Date(names(url_status))
+      nhs_recent_url <- try_urls[as.character(max(names(url_status)))]
+      # Get latest url download
       recent <- download_excel(nhs_recent_url,
-        "nhs_recent.xlsx",
-        verbose = self$verbose,
-        transpose = TRUE,
-        sheet = 1,
-        range = cell_limits(c(28, 2), c(36, NA))
-      )
+                               "nhs_recent.xlsx",
+                               verbose = self$verbose,
+                               transpose = TRUE,
+                               sheet = 1,
+                               range = cell_limits(c(28, 2), c(36, NA))
+        )
 
       # 2. Data for August 2020 to 7 April 2021
       archive <- download_excel(
