+++ NAMESPACE
@@ -74,8 +74,10 @@ importFrom(dplyr,summarise)
 importFrom(dplyr,tally)
 importFrom(dplyr,ungroup)
 importFrom(dplyr,vars)
+importFrom(httr,GET)
 importFrom(httr,POST)
 importFrom(httr,content)
+importFrom(httr,status_code)
 importFrom(lifecycle,deprecate_warn)
 importFrom(lifecycle,deprecated)
 importFrom(lifecycle,is_present)
@@ -91,6 +93,7 @@ importFrom(memoise,cache_filesystem)
 importFrom(memoise,memoise)
 importFrom(purrr,compact)
 importFrom(purrr,map)
+importFrom(purrr,map_chr)
 importFrom(purrr,map_lgl)
 importFrom(purrr,reduce)
 importFrom(purrr,safely)
