### 2021-06-04 e2e063f
Update NEWS.md
