+++ pkgdown.yml
@@ -1,12 +1,12 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 71a93f1d9b3ccb66dcdd3bc0b690b4e7dd55779c
+pkgdown_sha: 900ef5471576f8f29508a7e7d94fae51390f2d96
 articles:
   dataset-status: dataset-status.html
   quickstart: quickstart.html
   slowstart: slowstart.html
   testing: testing.html
-last_built: 2021-06-04T01:47Z
+last_built: 2021-06-04T16:35Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
