### 2021-06-07 0801639
Adjust to only be run on regional type datasets
