+++ README.Rmd
@@ -28,11 +28,10 @@ install.packages("covidregionaldata")
 ```
 Install the stable development version of the package with:
 
-```{r, eval = FALSE}
-install.packages("drat")
-drat:::add("epiforecasts")
-install.packages("covidregionaldata")
+```{r, eval = TRUE}
+install.packages("covidregionaldata", repos = "https://epiforecasts.r-universe.dev")
 ```
+
 
 Install the unstable development version of the package with:
 
