### 2021-06-10 a7fbe00
added CMMID working group mention to readme
