+++ README.md
@@ -34,6 +34,15 @@ Wherever available data is provided on deaths, hospitalisations, and
 tests. National level data is also supported using a range of data
 sources as well as line list data and links to intervention data sets.
 
+## Supported countries
+
+This map shows what countries have level 1 and level 2 subregion data
+directly from an official source within that country. Please note other
+countries may be provided through our interface to external data
+sources, such as WHO and JHU.
+
+<img src="man/figures/README-supported_region_plot-1.png" width="100%" />
+
 ## Installation
 
 Install from CRAN:
@@ -46,6 +55,8 @@ Install the stable development version of the package with:
 
 ``` r
 install.packages("covidregionaldata", repos = "https://epiforecasts.r-universe.dev")
+#> Installing package into '/home/joe/R/x86_64-pc-linux-gnu-library/3.6'
+#> (as 'lib' is unspecified)
 ```
 
 Install the unstable development version of the package with:
@@ -76,7 +87,7 @@ the temporary directory by default),
 
 ``` r
 start_using_memoise()
-#> Using a cache at: /tmp/RtmpcQYGco
+#> Using a cache at: /tmp/RtmpTq89jF
 ```
 
 To stop using `memoise` use,
@@ -101,7 +112,7 @@ the Google COVID-19 open data project), use:
 ``` r
 nots <- get_national_data()
 #> Downloading data from https://covid19.who.int/WHO-COVID-19-global-data.csv
-#> Rows: 123,003
+#> Rows: 124,188
 #> Columns: 8
 #> Delimiter: ","
 #> chr  [3]: Country_code, Country, WHO_region
@@ -113,7 +124,7 @@ nots <- get_national_data()
 #> Cleaning data
 #> Processing data
 nots
-#> # A tibble: 123,003 x 15
+#> # A tibble: 124,188 x 15
 #>    date       un_region who_region country        iso_code cases_new cases_total
 #>    <date>     <chr>     <chr>      <chr>          <chr>        <dbl>       <dbl>
 #>  1 2020-01-03 Asia      EMRO       Afghanistan    AF               0           0
@@ -126,7 +137,7 @@ nots
 #>  8 2020-01-03 Americas  AMRO       Antigua & Bar… AG               0           0
 #>  9 2020-01-03 Americas  AMRO       Argentina      AR               0           0
 #> 10 2020-01-03 Asia      EURO       Armenia        AM               0           0
-#> # … with 122,993 more rows, and 8 more variables: deaths_new <dbl>,
+#> # … with 124,178 more rows, and 8 more variables: deaths_new <dbl>,
 #> #   deaths_total <dbl>, recovered_new <dbl>, recovered_total <dbl>,
 #> #   hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>, tested_total <dbl>
 ```
@@ -167,7 +178,7 @@ for example by level 1 region in the UK, use:
 ``` r
 uk_nots <- get_regional_data(country = "UK", verbose = FALSE)
 uk_nots
-#> # A tibble: 6,396 x 26
+#> # A tibble: 6,461 x 26
 #>    date       region   region_code cases_new cases_total deaths_new deaths_total
 #>    <date>     <chr>    <chr>           <dbl>       <dbl>      <dbl>        <dbl>
 #>  1 2020-01-30 East Mi… E12000004          NA          NA         NA           NA
@@ -180,7 +191,7 @@ uk_nots
 #>  8 2020-01-30 Scotland S92000003          NA          NA         NA           NA
 #>  9 2020-01-30 South E… E12000008          NA          NA         NA           NA
[…]
