### 2021-06-10 60469fc
made recomended changes to plot code
