+++ inst/paper/paper.md
@@ -58,11 +58,11 @@ The onset of the COVID-19 pandemic in late 2019 has placed pressure on public he
 
 At a national level, epidemiological COVID-19 data is available to download from official sources such as the [World Health Organisation (WHO)](https://covid19.who.int/) [@WorldHealthOrganisation] or the [European Centre for Disease Control (ECDC)](https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide%7D) [@EuropeanCentreforDiseasePreventionandControl]. Many government bodies provide a wider range of country specific data, such as [Public Health England in the United Kingdom](https://coronavirus.data.gov.uk/details/about-data) [@PublicHealthEngland], and this is often the only way to access data at a subnational scale, for example by state, district, or province.
 
-Sometimes collated from a range of national and subnational sources, these data come in a variety of formats, requiring users to check and standardise data before it can be combined or processed for analysis.  This is a particularly time-consuming process for subnational data sets, which are often only available in the originating countries’ languages and require customised methods for downloading and processing. This generates potential for errors through programming mistakes, changes to functions defined in other packages, or unexpected changes to a data source. This can lead to misrepresenting the data in ways which are difficult to identify. At best, an independent data processing workflow only slows down the pace of research and analysis, while at worst it can lead to misleading and erroneous results.
+Sometimes collated from a range of national and subnational sources, these data come in a variety of formats, requiring users to check and standardise data before it can be combined or processed for analysis.  This is a particularly time-consuming process for subnational data sets, which are often only available in the originating countries’ languages and require customised methods for downloading and processing. This generates potential for errors through programming mistakes, changes to a dependency package, or unexpected changes to a data source. This can lead to misrepresenting the data in ways which are difficult to identify. At best, an independent data processing workflow only slows down the pace of research and analysis, while at worst it can lead to misleading and erroneous results.
 
 Because of these issues, it is important to develop robust tools that provide cleaned, checked and standardised data from multiple sources in a transparent manner. `covidregionaldata` provides easy access to clean data using a single-argument function, ready for analysing the epidemiology of COVID-19 from local to global scales, and in a framework that is easy to trace from raw data to the final standardised data set. Additional arguments to this function support users to, amongst other options, specify the spatial level of subnational data, return data with either standardised or country-specific variable names, or to access the full pipeline from raw to clean data. By default, cleaned and processed data is returned, however, the raw data from a source can also be returned. All data sources are checked daily via Github workflows and their status reported in the documentation section 'Data Status'. `covidregionaldata` largely depends on popular packages that many researchers are familiar with (such as the `tidyverse` suite [@Wickham2019]) and can therefore be easily adopted by researchers working in R. In addition to code coverage tests, we test and report the status of all data sets daily.
 
-Currently, `covidregionaldata` provides subnational data collated by official government bodies or by credible non-governmental efforts for 14 countries, including the UK, India, USA, and Brazil. It also provides an interface to subnational data curated by Johns Hopkins University [@Dong2020], and the [Google COVID-19 open data project](https://github.com/GoogleCloudPlatform/covid-19-open-data) [@Wahltinez2020]. National-level data is provided from the World Health Organisation (WHO) [@WorldHealthOrganisation], European Centre for Disease Prevention and Control (ECDC) [@EuropeanCentreforDiseasePreventionandControl], Johns Hopkins University (JHU) [@Dong2020], and the Google COVID-19 open data project [@Wahltinez2020].
[…]
