+++ README.md
@@ -22,9 +22,15 @@ commits](https://img.shields.io/github/commits-since/epiforecasts/covidregionald
 [![DOI](https://zenodo.org/badge/271601189.svg)](https://zenodo.org/badge/latestdoi/271601189)
 [![status](https://joss.theoj.org/papers/dd6f7acdae3b7136a3ac373ce9a0655c/status.svg)](https://joss.theoj.org/papers/dd6f7acdae3b7136a3ac373ce9a0655c)
 
-An interface to subnational and national level COVID-19 data. For all
+Interface to subnational and national level COVID-19 data sourced from
+both official sources, such as Public Health England in the UK, and from
+other Covid-19 data collections, including the World Health Organisation
+(WHO), European Centre for Disease Prevention and Control (ECDC), John
+Hopkins University (JHU), Google Open Data and others. Covid-19 data is
+cleaned and processed from their raw format in an open and transparent
+way, allowing users to scrutinise and extend our methods. For all
 countries supported, this includes a daily time-series of cases.
-Wherever available we also provide data on deaths, hospitalisations, and
+Wherever available data is provided on deaths, hospitalisations, and
 tests. National level data is also supported using a range of data
 sources as well as line list data and links to intervention data sets.
 This package is designed for people who wan’t access to standardised
@@ -42,8 +48,6 @@ Install the stable development version of the package with:
 
 ``` r
 install.packages("covidregionaldata", repos = "https://epiforecasts.r-universe.dev")
-#> Installing package into '/home/joe/R/x86_64-pc-linux-gnu-library/3.6'
-#> (as 'lib' is unspecified)
 ```
 
 Install the unstable development version of the package with:
@@ -74,7 +78,7 @@ the temporary directory by default),
 
 ``` r
 start_using_memoise()
-#> Using a cache at: /tmp/RtmppJ4l0P
+#> Using a cache at: /tmp/RtmpLzqu5D
 ```
 
 To stop using `memoise` use,
@@ -99,23 +103,34 @@ the Google COVID-19 open data project), use:
 ``` r
 nots <- get_national_data()
 #> Downloading data from https://covid19.who.int/WHO-COVID-19-global-data.csv
+#> Rows: 124,313
+#> Columns: 8
+#> Delimiter: ","
+#> chr  [3]: Country_code, Country, WHO_region
+#> dbl  [4]: New_cases, Cumulative_cases, New_deaths, Cumulative_deaths
+#> date [1]: Date_reported
+#> 
+#> Use `spec()` to retrieve the guessed column specification
+#> Pass a specification to the `col_types` argument to quiet this message
 #> Cleaning data
 #> Processing data
 nots
-#> # A tibble: 124,188 x 15
-#>    date       un_region who_region country           iso_code cases_new cases_total deaths_new deaths_total recovered_new recovered_total hosp_new hosp_total tested_new tested_total
-#>    <date>     <chr>     <chr>      <chr>             <chr>        <dbl>       <dbl>      <dbl>        <dbl>         <dbl>           <dbl>    <dbl>      <dbl>      <dbl>        <dbl>
-#>  1 2020-01-03 Asia      EMRO       Afghanistan       AF               0           0          0            0            NA              NA       NA         NA         NA           NA
-#>  2 2020-01-03 Europe    EURO       Albania           AL               0           0          0            0            NA              NA       NA         NA         NA           NA
-#>  3 2020-01-03 Africa    AFRO       Algeria           DZ               0           0          0            0            NA              NA       NA         NA         NA           NA
-#>  4 2020-01-03 Oceania   WPRO       American Samoa    AS               0           0          0            0            NA              NA       NA         NA         NA           NA
[…]

[…]
