+++ NEWS.md
@@ -1,4 +1,4 @@
-# covidregionaldata 0.9.2
+# covidregionaldata 0.9.3
 
 This release is in development and therefore new features may change without warning (though we will seek to keep this to a minimum).
 
@@ -8,10 +8,11 @@ This release is in development and therefore new features may change without war
 * Support added for data from the European Commission's Joint Research Centre (JRC). The source aggregates incidence data at the country and regional level for 34 UCPM Participating States plus Switzerland (by @joseph-palmer).
 * Support added for data from the Netherlands provided by RVIM (English: National Institute for Public Health and the Environment). This source provides case, deaths and hospital admission data at the province and municipal levels (by @joseph-palmer).
 * Support added for data from Switzerland and Liechtenstein collated by Canton Zurich (@OpenZH). This source provides case, deaths and hospital admission data at the canton level (by @RichardMN).
+* Made package changes recomended in the JOSS review, including additional statements of need to the README, updates to the manuscript (paper.md) and fixes a bug of multiple sources for some countries.
 
 ## Changes to implemented data sources
 
-* Increased the robustness of fetching UK NHS admissions by region. Rather than testing a single date for data we now look over the last 7 days and pick the most recent availabe data set (by @kathsherratt).
+* Increased the robustness of fetching UK NHS admissions by region. Rather than testing a single date for data we now look over the last 7 days and pick the most recent available data set (by @kathsherratt).
 
 ## Other changes
 
