+++ README.md
@@ -26,15 +26,16 @@ Interface to subnational and national level COVID-19 data sourced from
 both official sources, such as Public Health England in the UK, and from
 other Covid-19 data collections, including the World Health Organisation
 (WHO), European Centre for Disease Prevention and Control (ECDC), John
-Hopkins University (JHU), Google Open Data and others. Covid-19 data is
-cleaned and processed from their raw format in an open and transparent
-way, allowing users to scrutinise and extend our methods. For all
-countries supported, this includes a daily time-series of cases.
-Wherever available data is provided on deaths, hospitalisations, and
-tests. National level data is also supported using a range of data
-sources as well as line list data and links to intervention data sets.
-This package is designed for people who wan’t access to standardised
-Covid-19 data from ‘official’ sources.
+Hopkins University (JHU), Google Open Data and others. Many Covid-19
+data libraries provide an interface to a maintained database, this
+package is designed to take Covid-19 data from different sources and
+clean/process it from its raw format in an open and transparent way This
+allows users to inspect and scrutinise where their data comes from and
+how it is standardised. For all countries supported, the data includes a
+daily time-series of cases and, wherever available, data on deaths,
+hospitalisations, and tests. National level data is also supported using
+a range of data sources as well as line list data and links to
+intervention data sets.
 
 ## Installation
 
@@ -78,7 +79,7 @@ the temporary directory by default),
 
 ``` r
 start_using_memoise()
-#> Using a cache at: /tmp/RtmpLzqu5D
+#> Using a cache at: /tmp/RtmpGsHq5i
 ```
 
 To stop using `memoise` use,
@@ -103,7 +104,7 @@ the Google COVID-19 open data project), use:
 ``` r
 nots <- get_national_data()
 #> Downloading data from https://covid19.who.int/WHO-COVID-19-global-data.csv
-#> Rows: 124,313
+#> Rows: 124,425
 #> Columns: 8
 #> Delimiter: ","
 #> chr  [3]: Country_code, Country, WHO_region
@@ -115,7 +116,7 @@ nots <- get_national_data()
 #> Cleaning data
 #> Processing data
 nots
-#> # A tibble: 124,424 x 15
+#> # A tibble: 124,425 x 15
 #>    date       un_region who_region country        iso_code cases_new cases_total
 #>    <date>     <chr>     <chr>      <chr>          <chr>        <dbl>       <dbl>
 #>  1 2020-01-03 Asia      EMRO       Afghanistan    AF               0           0
@@ -128,7 +129,7 @@ nots
 #>  8 2020-01-03 Americas  AMRO       Antigua & Bar… AG               0           0
 #>  9 2020-01-03 Americas  AMRO       Argentina      AR               0           0
 #> 10 2020-01-03 Asia      EURO       Armenia        AM               0           0
-#> # … with 124,414 more rows, and 8 more variables: deaths_new <dbl>,
+#> # … with 124,415 more rows, and 8 more variables: deaths_new <dbl>,
 #> #   deaths_total <dbl>, recovered_new <dbl>, recovered_total <dbl>,
 #> #   hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>, tested_total <dbl>
 ```
@@ -169,7 +170,7 @@ for example by level 1 region in the UK, use:
 ``` r
 uk_nots <- get_regional_data(country = "UK", verbose = FALSE)
 uk_nots
-#> # A tibble: 6,461 x 26
+#> # A tibble: 6,474 x 26
 #>    date       region   region_code cases_new cases_total deaths_new deaths_total
[…]
