### 2021-06-11 1927a81
updated to run the chunks and load tables from the code
