+++ vignettes/slowstart.Rmd
@@ -29,7 +29,7 @@ library(covidregionaldata)
 
 Both the World Health Organisation (WHO) and European Centre for Disease Control (ECDC) provide worldwide national data. Access national level data for any country using:
 
-```{r}
+```{r, eval=FALSE}
 get_national_data()
 ```
 
@@ -47,24 +47,17 @@ This function returns daily new and cumulative (total) cases, and where availabl
 
 As for national level data any gaps in reported data are filled with NAs.
 
-For example, data for Belgium Level 1 regions over time can be accessed using:
+For example, data for Italy Level 1 regions over time can be accessed using:
 
-```{r}
-get_regional_data(countries = "Belgium")
+```{r, eval=TRUE, message=FALSE}
+get_regional_data(country = "france")
 ```
-This returns a dataset in this format:
-
-**date**|**region**|**iso\_code**|**cases\_new**|**cases\_total**|**deaths\_new**|**deaths\_total**|**recovered\_new**|**recovered\_total**|**hosp\_new**|**hosp\_total**|**tested\_new**|**tested\_total**
-:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:
-2020-05-24|Wallonia|BE-WAL|24|18196|16|3251|NA|NA|8|5126|NA|NA
-2020-05-25|Brussels|BE-BRU|26|5838|2|1421|NA|NA|6|2533|NA|NA
-2020-05-25|Flanders|BE-VLG|183|32381|14|4681|NA|NA|29|9334|NA|NA
 
 Alternatively, the same data can be accessed using the underlying class as follows (the belgium object now contains data at each processing step and the methods used at each step), 
 
-```{r}
-belgium <- Belgium$new(get = TRUE)
-belgium$return()
+```{r, eval=TRUE, message=FALSE}
+italy <- Italy$new(get = TRUE)
+italy$return()
 ```
 ### Level 1 and Level 2 regions
 
@@ -74,31 +67,19 @@ Data for Level 2 units can be returned by using the `level = "2"` argument. The 
 
 An example of a country with Level 2 units is Belgium, where Level 2 units are Belgian provinces:
 
-```{r}
-get_regional_data("Belgium", level = "2")
+```{r, eval=TRUE, message=FALSE}
+get_regional_data(country = "france", level = "2")
 ```
-This returns a dataset with the format:
-
-**date**|**province**|**level\_2\_region\_code**|**region**|**iso\_code**|**cases\_new**|**cases\_total**|**deaths\_new**|**deaths\_total**|**recovered\_new**|**recovered\_total**|**hosp\_new**|**hosp\_total**|**tested\_new**|**tested\_total**
-:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:
-2020-05-24|Brussels|BE-BRU|Brussels|BE-BRU|7|5812|NA|NA|NA|NA|4|2527|NA|NA
-2020-05-24|Antwerpen|BE-VAN|Flanders|BE-VLG|16|7905|NA|NA|NA|NA|5|2510|NA|NA
-2020-05-24|Limburg|BE-VLI|Flanders|BE-VLG|14|6126|NA|NA|NA|NA|2|1848|NA|NA
 
 ### Totals
 
 For totalled data up to the most recent date available, use the `totals` argument.
 
 ```{r}
-get_regional_data("Belgium", totals = TRUE)
+get_regional_data("france", totals = TRUE)
 ```
[…]
