+++ vignettes/slowstart.Rmd
@@ -19,7 +19,7 @@ knitr::opts_chunk$set(
 
 The package can either be installed from CRAN, from our `drat` repository, or from GitHub. See the README for details. Once installed load the package using the following,
 
-```{r}
+```{r, eval=TRUE}
 library(covidregionaldata)
 ```
 
