### 2021-06-11 ec635a6
updated vignette title and tried to fix check error
