+++ DESCRIPTION
@@ -108,7 +108,6 @@ Suggests:
     rmarkdown,
     rvest,
     rworldmap,
-    sf,
     spelling,
     testthat (>= 3.0.0),
     usethis
