+++ vignettes/slowstart.Rmd
@@ -47,17 +47,22 @@ This function returns daily new and cumulative (total) cases, and where availabl
 
 As for national level data any gaps in reported data are filled with NAs.
 
-For example, data for Italy Level 1 regions over time can be accessed using:
+For example, data for France Level 1 regions over time can be accessed using:
 
-```{r, eval=TRUE, message=FALSE}
+```{r}
 get_regional_data(country = "france")
 ```
+```{r, echo=FALSE, eval=TRUE, message=FALSE}
+knitr::kable(
+  tail(get_regional_data(country = "france"), n = 5)
+)
+```
 
-Alternatively, the same data can be accessed using the underlying class as follows (the belgium object now contains data at each processing step and the methods used at each step), 
+Alternatively, the same data can be accessed using the underlying class as follows (the France object now contains data at each processing step and the methods used at each step), 
 
-```{r, eval=TRUE, message=FALSE}
-italy <- Italy$new(get = TRUE)
-italy$return()
+```{r, eval=FALSE, message=FALSE}
+france <- France$new(get = TRUE)
+france$return()
 ```
 ### Level 1 and Level 2 regions
 
@@ -65,10 +70,15 @@ All countries included in the package (see below,"Coverage") have data for regio
 
 Data for Level 2 units can be returned by using the `level = "2"` argument. The dataset will still show the corresponding Level 1 region.
 
-An example of a country with Level 2 units is Belgium, where Level 2 units are Belgian provinces:
+An example of a country with Level 2 units is France, where Level 2 units are French departments:
 
-```{r, eval=TRUE, message=FALSE}
+```{r}
 get_regional_data(country = "france", level = "2")
+```
+```{r, echo=FALSE, eval=TRUE, message=FALSE}
+knitr::kable(
+  tail(get_regional_data(country = "france", level = "2"), n = 5)
+)
 ```
 
 ### Totals
@@ -77,6 +87,11 @@ For totalled data up to the most recent date available, use the `totals` argumen
 
 ```{r}
 get_regional_data("france", totals = TRUE)
+```
+```{r, echo=FALSE, eval=TRUE, message=FALSE}
+knitr::kable(
+  tail(get_regional_data(country = "france", totals = TRUE), n = 5)
+)
 ```
 
 
