### 2021-06-11 e9f8e6c
Update mention of drat to recommend r-universe instead
Follow up of #376
