### 2021-06-11 a2bc58f
Fix argument name in vignette
