### 2021-06-11 07bef36
updated vignette name, pkgsite and link in readme
