+++ DESCRIPTION
@@ -108,6 +108,7 @@ Suggests:
     rmarkdown,
     rvest,
     rworldmap,
+    sf,
     spelling,
     testthat (>= 3.0.0),
     usethis
