+++ README.Rmd
@@ -149,7 +149,7 @@ citation("covidregionaldata")
 
 [![Development](https://img.shields.io/badge/Wiki-lightblue.svg?style=flat)](https://github.com/epiforecasts/covidregionaldata/wiki/)
 
-This package is the result of hard work from a number of contributors (see contributors list in the DESCRIPTION). We would like to thank the [CMMID COVID-19 working group
+This package is the result of work from a number of contributors (see contributors list [here](https://epiforecasts.io/covidregionaldata/authors.html)). We would like to thank the [CMMID COVID-19 working group
 ](https://cmmid.github.io/groups/ncov-group.html) for inciteful comments and feedback.
 
 We welcome contributions and new contributors! We particularly appreciate help adding new data sources for countries at sub-national level, or work on priority problems in the [issues](https://github.com/epiforecasts/covidregionaldata/issues). Please check and add to the issues, and/or add a [pull request](https://github.com/epiforecasts/covidregionaldata/pulls). For more details, start with the [contributing guide](https://github.com/epiforecasts/covidregionaldata/wiki/Contributing). For details of the steps required to add support for a dataset see the [adding data guide](https://github.com/epiforecasts/covidregionaldata/wiki/Adding-Data).
