+++ DESCRIPTION
@@ -2,20 +2,16 @@ Package: covidregionaldata
 Title: Subnational Data for the COVID-19 Outbreak
 Version: 0.9.2.4000
 Authors@R:
-    c(person(given = "Sam Abbott",
-             role = c("aut", "cre"),
-             email = "sam.abbott@lshtm.ac.uk",
-             comment = c(ORCID = "0000-0001-8057-8037")),
+    c(person(given = "Joe",
+             family = "Palmer",
+             role = "aut",
+             email = "joe.palmer.2019@live.rhul.ac.uk",
+             comment = c(ORCID = "0000-0002-5593-9352")),
       person(given = "Katharine",
              family = "Sherratt",
              role = "aut",
              email = "katharine.sherratt@lshtm.ac.uk",
              comment = c(ORCID = "0000-0003-2049-3423")),
-      person(given = "Joe",
-             family = "Palmer",
-             role = "aut",
-             email = "joe.palmer.2019@live.rhul.ac.uk",
-             comment = c(ORCID = "0000-0002-5593-9352")),
       person(given = "Richard",
              family = "Martin-Nielsen",
              role = "aut",
@@ -62,19 +58,24 @@ Authors@R:
       person(given = "Sebastian",
              family = "Funk",
              role = "aut",
-             email = "sebastian.funk@lshtm.ac.uk"))
+             email = "sebastian.funk@lshtm.ac.uk"),
+      person(given = "Sam Abbott",
+             role = c("aut", "cre"),
+             email = "sam.abbott@lshtm.ac.uk",
+             comment = c(ORCID = "0000-0001-8057-8037")))
 Description: An interface to subnational and national level COVID-19 data
     sourced from both official sources, such as Public Health England in
     the UK, and from other Covid-19 data collections, including the World
     Health Organisation (WHO), European Centre for Disease Prevention and
     Control (ECDC), John Hopkins University (JHU), Google Open Data and
-    others. Facilitates cleaning and processing Covid-19 data their raw 
-    format in an open and transparent way, allowing users to scrutinise and
-    extend the provided methods. For all countries supported, data includes
-    a daily time-series of cases. Wherever available data is also provided
-    for deaths, hospitalisations, and tests. National level data are also
-    supported using a range of sources as well as line list data and
-    links to intervention data sets.
+    others. Designed to streamline Covid-19 data extraction, cleaning,
+    and processing from a range of data sources in an open and transparent
+    way. This allows users to inspect and scrutinise the data, and tools
+    used to process it, at every step. For all countries supported, data
+    includes a daily time-series of cases. Wherever available data is also
+    provided for deaths, hospitalisations, and tests. National level data
+    are also supported using a range of sources as well as line list data
+    and links to intervention data sets.
 License: MIT + file LICENSE
 URL: https://epiforecasts.io/covidregionaldata/,
     https://github.com/epiforecasts/covidregionaldata/
