+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: covidregionaldata
 Title: Subnational Data for the COVID-19 Outbreak
-Version: 0.9.2.4000
+Version: 0.9.2.5000
 Authors@R:
     c(person(given = "Joe",
              family = "Palmer",
