+++ pkgdown.yml
@@ -2,11 +2,11 @@ pandoc: 2.7.3
 pkgdown: 1.6.1.9001
 pkgdown_sha: 9ce6af43b55cd21e9563526e12a12d1243657bf6
 articles:
-  dataset-status: dataset-status.html
   quickstart: quickstart.html
   slowstart: slowstart.html
+  supported-countries: supported-countries.html
   testing: testing.html
-last_built: 2021-06-14T10:48Z
+last_built: 2021-06-14T13:58Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
