+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   slowstart: slowstart.html
   supported-countries: supported-countries.html
   testing: testing.html
-last_built: 2021-06-14T13:58Z
+last_built: 2021-06-14T14:10Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
