### 2021-06-14 #3290 @joseph-palmer
@federicomarini

> maybe use a newer reference to the R core team?

We have updated the reference in the manuscript.

> l 16-17, repetition of extensible/extend

We have changed this in the manuscript

> l.38: "changes to a dependency package" is not so clearly defined, do you mean something on the line of "changes in the subsequent processing steps" (for packages that use these sources)?

What we mean by this is if a function used from another package (e.g. dplyr, a package on which this package depends on) changes then the function may not work the same way. As far we are aware this is standard terminology.

> l.53: "test and report... daily" -> You could specify this happens via GHA and their output is recorded on the project webpage

Certainly, I have updated the manuscript to include this.

> while reading l.42-47: as a general idea: is there some way to track/record the origin or snapshot of the data for the files that are sourced? Something like an attribute could do it, but maybe some similar mechanism is already in place

This is covered in the documentation, all the data is stored in the country class in the ‘data’ field, which contains ‘raw’ data (as is when downloaded) in addition to cleaned and processed data.

> Are the versions for sars2pack and covdata needed to ensure that this is the status quo at the moment of writing the manuscript? If dois are available, that could be even better

The citations for these are taken from the package using the ‘citation()’ tool in R so are just as the authors report them.

> Footnote: is it correct to have a $ after Palmer as in Palmer$ et al? Maybe it is a feature of the build system

No this is an error, good spot! I have updated this in the manuscript.

> A statement of need: These are elaborated in the vignettes. Maybe add a (more explicit) statement regarding what the target audience would be?

I have expanded on this point in the README but have stopped short of an explicit statement to the tune of “this package is designed for ….”. This is just personal preference and I think the contents of the README now, along with the other documentation provides potential users with enough information to decide if this package is right for them.

> Installation instructions: The dependencies are all installed and correctly handled - maybe you can point out that to run the vignette examples you might need to use dependencies = TRUE in the installation command?

We would generally advise individuals to check out the pkgdown documentation rather than build the vignette but we agree that it should be easy for users to run in R. As far as we are aware having library calls in the vignettes is standard practice for suggested packages required to build them that are not included as package imports. This would naturally support a workflow where the less technical user (who may not know about suggested packages) installs the packages that are not in place when they go to build the vignette.

> Example usage: The vignettes are present, and a very welcome piece of documentation. It might be overkill, but maybe some "fully worked out use cases" in one of the studies where the package has been used can add even more value? I realize there might be also some runtime/check-time constraints, so it is up to the authors to explore this point.

Absolutely, we plan to expand the documentation vignettes as we move forwards with the package (https://github.com/epiforecasts/covidregionaldata/issues/316). Hopefully, you agree that the documentation as it stands is sufficient at this time until we can spare development time for additional documentation?

> Automated tests: The test suite is well structured, and covers a wide portion of the code to retrieve and process the data. Maybe some additional tests might be desired to increase the overall coverage? Again, it might be a tradeoff between test runtimes and package coverage…

Completely agree and we have currently open PRs to increase test coverage (https://github.com/epiforecasts/covidregionaldata/pull/312) (https://github.com/epiforecasts/covidregionaldata/pull/307). We feel that the current testing framework probably has greater coverage than comparable tools but we are very keen to develop this further.

> the help in some classes seems redundant, see ?Belgium (multiple same urls)

Thank you for spotting this, it somehow slipped by us! This was an error and we have now corrected this for Belgium and others with this bug.

> maybe have the slowstart vignette evaluated? By having these chunks automatically generated, it would be robust to changes in the underlying data

This is a great idea and we have updated the vignette to run these chunks and use the code to render the tables. Complying with CRAN policy may be an issue here but if at submission time for this version we have issue we will alter the vignette processing pipeline so that CRAN is passed an md version of this Rmd but that we otherwise use an Rmd.
