### 2021-06-14 #3290 @joseph-palmer
@mponce0

> I'm still not quite sure about this, I understand your reasoning but I'm not sure the version of the package must be tied to the data being incorporated... this will generate several versions as contributors start to add more and more data... I'm not saying I'm against just doubting here...

I understand your concerns, however, versioning for most purposes just needs to be a way of marking a difference between releases. For our project, as new datasets get added we then issue a new release with these datasets (usually as a batch) to CRAN, so there must be some version difference. Having versioning take account of the data included suits our approach.

> Also this makes me think about what would happen when one of the data sources is for whatever reason not available anymore... will you prune it or perhaps an alternative approach is to create a copy of the original dataset and keep as "legacy" data? I'm still unsure about the versioning...

At the moment we do not store data snapshots as the package is designed to be purely an interface to existing data sources. Should the datasets change source we would aim to update our package to use the new source instead of the old. If the source disappeared we would try and find the legacy source and offer this as an option, however it is not within the scope of this package to store data. An option we are considering is hosting a separate legacy source to use in these circumstances but we have not yet made much progress with these plans.

> Sure, I know that and I checked it, but this is requested from JOSS... if JOSS ( @csoneson ) is OK with it I don't have an issue... just following their guidelines. For instance, I have added a short paragraph in the web README in my own packages due to this request, just FYI.

Yes I see this is not as clear as I thought, I have added more details of the working group to the README document. Your example was very helpful to follow, thank you!

>OK, but in your README file I'm still seeing "R-CMD-check" Failing! is this because one of the data pipelines was failing or something else? In any case, it should be unentangled from the data verification process.

Data testing is split from other unit testing. By this we mean that downloading data is checked along with its integrity on a scheduled CI check each night. Our general check (and what is shown on the front page of the package) is a more traditional check that uses stored data to  ensure the internal consistency of the package. This was a crucial design element as testing external data sources is error prone and would likely lead to our package not being CRAN stable at the very least.  We had a general CI outage across projects that is now resolved and may have been what you saw.

> But what would happen to users who have already installed the package and try to access a dataset which is not longer available or failing? It will fail, right? No issues here but just be sure to provide an informative message to the user, recall that many R users (most probably) are not so "savvy" to be involved with dev versions, git-repos, etc.

Yes, if a source is no longer supported we would provide a depreciation message but the user would need to install the latest version to get this. If using an older version which had code to collect data that no longer existed then this would just fail. Currently we only support data status updates via a vignette (though this is available in the downloaded version of the package) we are also considering if surfacing a user function for this check makes sense. We have recently had user feedback that the data status page is useful and currently no user feedback along the lines of is my data broken etc. which we had prior to the data status page being in place. Very much agree that this is something it is important to make easier to check so we will keep an eye on user feedback to see if we need to make changes.
