### 2021-06-14 #3290 @joseph-palmer
@csoneson

> On the software side itself I think the first step is indeed to include the contributors in the DESCRIPTION file together with the type of contribution, as is currently done. Any additional clarifications, acknowledgements etc would, in my opinion, fit well in the README.

Certainly, we have updated these to ensure it meets the guidance.

> I had just one question regarding the working group: you say above that it was included since "members of the group made sizeable contributions to the planning and design of the package". I also note that several of the named authors are indeed members of this working group. Could you clarify whether there were additional contributions made by the working group (beyond the named individuals)?

Sure, individual members of the working group made very specific contributions hence their inclusion in the description and git commit histories. The package was discussed in the working group meetings at great length and whilst no code was actually written in these meetings and by some members, we wanted to acknowledge the role the working group had as a whole, including contributors listed and those not, in helping develop this package. I have included a statement in the README citing the working group along with a link to their site outlining what and who they are. Our inclusion of the working group is in line with our COVID-19 response group policy with the ideal of reflecting unstated contributions (such as finding data sources), often from those at an early career stage for whom it is most important, that would otherwise not be acknowledged.
