+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   slowstart: slowstart.html
   supported-countries: supported-countries.html
   testing: testing.html
-last_built: 2021-06-14T16:41Z
+last_built: 2021-06-15T01:03Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
