### 2021-06-15 3559d7e
made available_datasets a package dataset rather than something rendered every time
