+++ man/initialise_dataclass.Rd
@@ -111,6 +111,7 @@ Data interface functions
 \code{\link{DataClass}},
 \code{\link{get_available_datasets}()},
 \code{\link{get_national_data}()},
-\code{\link{get_regional_data}()}
+\code{\link{get_regional_data}()},
+\code{\link{render_available_datasets}()}
 }
 \concept{interface}
