+++ man/all_country_data.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datasets.R
+\docType{data}
+\name{all_country_data}
+\alias{all_country_data}
+\title{Table of available datasets along with level and other information.
+Rendered from the individual R6 class objects included in this package.}
+\format{
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 8 columns.
+}
+\usage{
+all_country_data
+}
+\value{
+A tibble of available datasets and related information.
+}
+\description{
+Available datasets
+}
+\keyword{datasets}
