+++ NAMESPACE
@@ -34,6 +34,7 @@ export(get_regional_data)
 export(initialise_dataclass)
 export(make_github_workflow)
 export(make_new_data_source)
+export(render_available_datasets)
 export(reset_cache)
 export(start_using_memoise)
 export(stop_using_memoise)
