### 2021-06-15 8105d64
added test to check all_data is up to date with all classes
