### 2021-06-15 512ed83
removed render_available_datasets, placed code in get_available__datasets with options for rendering. also added parameter for namespace and updated tests
