+++ man/get_national_data.Rd
@@ -100,7 +100,6 @@ Data interface functions
 \code{\link{DataClass}},
 \code{\link{get_available_datasets}()},
 \code{\link{get_regional_data}()},
-\code{\link{initialise_dataclass}()},
-\code{\link{render_available_datasets}()}
+\code{\link{initialise_dataclass}()}
 }
 \concept{interface}
