+++ man/get_regional_data.Rd
@@ -98,7 +98,6 @@ Data interface functions
 \code{\link{DataClass}},
 \code{\link{get_available_datasets}()},
 \code{\link{get_national_data}()},
-\code{\link{initialise_dataclass}()},
-\code{\link{render_available_datasets}()}
+\code{\link{initialise_dataclass}()}
 }
 \concept{interface}
