+++ man/get_available_datasets.Rd
@@ -2,9 +2,9 @@
 % Please edit documentation in R/get_available_datasets.R
 \name{get_available_datasets}
 \alias{get_available_datasets}
-\title{Fetch table of supported data sets}
+\title{Fetch of build table of supported data sets}
 \usage{
-get_available_datasets(type)
+get_available_datasets(type, render = FALSE, namespace = "covidregionaldata")
 }
 \arguments{
 \item{type}{A character vector indicating the types of data to
@@ -12,6 +12,13 @@ return. Current options include "national" (which are datasets at the
 national level which inherit from \code{CountryDataClass}) and
 "regional" (which are datasets at the regional level which inherit
 directly from \code{DataClass()}).}
+
+\item{render}{Bool. If TRUE the supported data set table is build from the
+available classes using \code{summary} methods. If FALSE (dafault) the supported
+data set table is taken from package data.}
+
+\item{namespace}{characterArray The name of the namespace to search for class
+objects. Defaults to 'covidregionaldata' as the package.}
 }
 \value{
 A list of available data sets and the spatial aggregation data is
@@ -19,8 +26,11 @@ available for.
 }
 \description{
 The function returns data on what countries are available from
-the data provided with this package. This is rendered through the function
-\code{render_available_datasets()}.
+the data provided with this package. If render = TRUE The function searches
+the environment for R6 class objects and extracts the summary information
+from the available classes using their \code{summary} methods. If render = FALSE
+the function will fetch the data from package data. In practice this
+means that it can be used to indicate supported data sets.
 }
 \examples{
 # see all available datasets
@@ -31,6 +41,9 @@ get_available_datasets("national")
 
 # see only regional level datasets
 get_available_datasets("regional")
+
+# render the data
+get_availdable_datasets(render = TRUE)
 }
 \seealso{
 Data interface functions
@@ -38,7 +51,6 @@ Data interface functions
 \code{\link{DataClass}},
 \code{\link{get_national_data}()},
 \code{\link{get_regional_data}()},
-\code{\link{initialise_dataclass}()},
-\code{\link{render_available_datasets}()}
+\code{\link{initialise_dataclass}()}
 }
 \concept{interface}
