+++ man/get_available_datasets.Rd
@@ -43,7 +43,7 @@ get_available_datasets("national")
 get_available_datasets("regional")
 
 # render the data
-get_availdable_datasets(render = TRUE)
+get_available_datasets(render = TRUE)
 }
 \seealso{
 Data interface functions
