### 2021-06-15 #3290 @csoneson
Thanks for this @joseph-palmer! This sounds sensible to me.

@mponce0, @federicomarini - could you take a look when you have a chance to see whether the responses above address your concerns? Thanks!
