+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   slowstart: slowstart.html
   supported-countries: supported-countries.html
   testing: testing.html
-last_built: 2021-06-15T01:03Z
+last_built: 2021-06-16T01:09Z
 urls:
   reference: epiforecasts.io/covidregionaldata/reference
   article: epiforecasts.io/covidregionaldata/articles
