+++ README.md
@@ -1,5 +1,5 @@
 
-# Subnational data for the COVID-19 outbreak
+# Subnational data for the COVID-19 outbreak <img src="man/figures/logo.png" align="right" alt="" width="120" />
 
 [![Lifecycle:
 maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html)
@@ -27,11 +27,11 @@ both official sources, such as Public Health England in the UK, and from
 other Covid-19 data collections, including the World Health Organisation
 (WHO), European Centre for Disease Prevention and Control (ECDC), John
 Hopkins University (JHU), Google Open Data and others. This package is
-designed to streamline Covid-19 data extraction, cleaning, and processing
-from a range of data sources in an open and transparent way. This allows
-users to inspect and scrutinise the data, and tools used to process it,
-at every step. For all countries supported, data includes a daily
-time-series of cases and, wherever available, data on deaths,
+designed to streamline Covid-19 data extraction, cleaning, and
+processing from a range of data sources in an open and transparent way.
+This allows users to inspect and scrutinise the data, and tools used to
+process it, at every step. For all countries supported, data includes a
+daily time-series of cases and, wherever available, data on deaths,
 hospitalisations, and tests. National level data is also supported using
 a range of data sources as well as line list data and links to
 intervention data sets.
@@ -78,7 +78,7 @@ the temporary directory by default),
 
 ``` r
 start_using_memoise()
-#> Using a cache at: /tmp/RtmpJAeC7J
+#> Using a cache at: /tmp/RtmprQRl5f
 ```
 
 To stop using `memoise` use,
@@ -103,7 +103,7 @@ the Google COVID-19 open data project), use:
 ``` r
 nots <- get_national_data()
 #> Downloading data from https://covid19.who.int/WHO-COVID-19-global-data.csv
-#> Rows: 124,425
+#> Rows: 125,724
 #> Columns: 8
 #> Delimiter: ","
 #> chr  [3]: Country_code, Country, WHO_region
@@ -115,7 +115,7 @@ nots <- get_national_data()
 #> Cleaning data
 #> Processing data
 nots
-#> # A tibble: 124,425 x 15
+#> # A tibble: 125,846 x 15
 #>    date       un_region who_region country        iso_code cases_new cases_total
 #>    <date>     <chr>     <chr>      <chr>          <chr>        <dbl>       <dbl>
 #>  1 2020-01-03 Asia      EMRO       Afghanistan    AF               0           0
@@ -128,7 +128,7 @@ nots
 #>  8 2020-01-03 Americas  AMRO       Antigua & Bar… AG               0           0
 #>  9 2020-01-03 Americas  AMRO       Argentina      AR               0           0
 #> 10 2020-01-03 Asia      EURO       Armenia        AM               0           0
-#> # … with 124,415 more rows, and 8 more variables: deaths_new <dbl>,
+#> # … with 125,836 more rows, and 8 more variables: deaths_new <dbl>,
 #> #   deaths_total <dbl>, recovered_new <dbl>, recovered_total <dbl>,
 #> #   hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>, tested_total <dbl>
 ```
@@ -169,7 +169,7 @@ for example by level 1 region in the UK, use:
 ``` r
 uk_nots <- get_regional_data(country = "UK", verbose = FALSE)
 uk_nots
-#> # A tibble: 6,474 x 26
+#> # A tibble: 6,539 x 26
 #>    date       region   region_code cases_new cases_total deaths_new deaths_total
[…]
