+++ .pre-commit-config.yaml
@@ -43,7 +43,7 @@ repos:
     rev: v3.4.0
     hooks: 
     -   id: check-added-large-files
-        args: ['--maxkb=200']
+        args: ['--maxkb=600']
     -   id: end-of-file-fixer
         exclude: '\.Rd'
 -   repo: local
