### 2021-06-16 0564367
Revert "made draft hexsticker"
This reverts commit 1b8fe5b857122d2ce67bc077f96002f78f8058a2.
