### 2021-06-17 #3290 @seabbs
Thanks for the great review and detailed feedback @federicomarini - very much appreciated. For interest, this is the source we were basing the statement about Rmd -> md vignettes on: https://ropensci.org/blog/2019/12/08/precompute-vignettes/
