+++ README.Rmd
@@ -15,13 +15,14 @@ knitr::opts_chunk$set(
 # Hexsticker designs
 
 ### Design 1
-![Design 1](man/figures/logo.png)
+![Design 1](man/figures/logo1.png)
 
 ### Design 2
 ![design 2](man/figures/logo2.png)
 
 ### Design 3
 ![design 3](man/figures/logo3.png)
+
 
 # Subnational data for the COVID-19 outbreak
 
