### 2021-06-20 a137ddf
Initial work to show regions for countries where we have regional data
