### 2021-06-21 5a11397
Fixed separation of countries with level 1 and level 2 data
