+++ inst/make_hexsticker.R
@@ -19,30 +19,35 @@ regional_countries <- get_available_datasets() %>%
 regional_countries_l2 <- regional_countries %>%
   filter(!(is.na(.data$level_2_region)))
 
-# get world data
-world <- ms_simplify(spData::world %>%
-  st_as_sf(), keep = 0.04)
+regional_countries_l1 <- regional_countries %>%
+  filter(is.na(.data$level_2_region))
 
-regional_maps <- ms_simplify(ne_states(gsub(' \\(.*\\)', "", regional_countries$origin,perl=TRUE),
-                           returnclass = "sf"), keep = 0.04) %>%
-    mutate(
-      region_code = paste("Level", woe_id %% 7 + 3)
-    )
+# get world data
+world <- spData::world %>%
+  st_as_sf()
+
+# regional_maps <- ms_simplify(ne_states(gsub(' \\(.*\\)', "", regional_countries$origin,perl=TRUE),
+#                            returnclass = "sf"), keep = 0.04) %>%
+#     mutate(
+#       region_code = paste("Level", woe_id %% 7 + 3)
+#     )
 
 
 world_without_regions <- ne_countries(returnclass = "sf") %>%
   filter(sovereignt != "Antarctica")
 
-regional_maps <- ms_simplify(ne_states(gsub(' \\(.*\\)', "", regional_countries$origin,perl=TRUE),
-                           returnclass = "sf"), keep = 0.04) %>%
-    mutate(
-      #region_code = paste("Level", woe_id %% 7 + 3)
-      region_code = woe_id %% 7 +
-        if_else(admin %in%
-                  countryname(regional_countries_l2[["origin"]], destination = "country.name.en"),
-                0, 7)
-    )
+numberOfLevels <- 4
+regional_maps_l1 <- ms_simplify(
+  ne_states(gsub(' \\(.*\\)', "",
+                 regional_countries_l1$origin, perl=TRUE),returnclass = "sf") %>%
+    mutate( region_code = woe_id %% numberOfLevels), keep = 0.04) 
 
+regional_maps_l2 <- ms_simplify(
+  ne_states(gsub(' \\(.*\\)', "",
+                 regional_countries_l2$origin, perl=TRUE),returnclass = "sf") %>%
+    mutate( region_code = woe_id %% numberOfLevels + numberOfLevels + 1), keep = 0.04) 
+
+regional_maps <- bind_rows(regional_maps_l1, regional_maps_l2)
 
 # mark supported countries from the world data
 supported_countries <- world %>%
