### 2021-06-22 #3290 @mponce0
@joseph-palmer Thank you for the explanations and updates to the different documents in the package.
Just recall to update your citation information to point towards the JOSS record when you get your final DOI.

@csoneson I'm good with the status of the manuscript and checked all the items in the check list, please let me know if there is anything else needed from my side.
