### 2021-06-22 ce0214d
Outlines added to supported countries, palette adjusted, code reorganised slightly.
