### 2021-06-23 ea6a48e
Change outlines to be lines, reorder display, remove commented code
