### 2021-06-24 #3290 @csoneson
👋🏻 @joseph-palmer - I just opened a PR with a couple of minor fixes to reference formatting. Please check and merge if you agree (and ask whedon to regenerate the pdf). The only other thing I noticed is the citation to Abbott et al (Zenodo), which doesn't have the same title as the actual Zenodo record.

Then, the next steps are as follows:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
