### 2021-06-24 00fd919
Adjusted summary method, began adding source text and url fields
