+++ R/Cuba.R
@@ -34,6 +34,12 @@ Cuba <- R6::R6Class("Cuba",
     # nolint end
     #' @field source_data_cols existing columns within the raw data
     source_data_cols = c("cases_new"),
+    #' @field source_text Plain text description of the source of the data
+    source_text = "COVID19 Cuba Data team",
+    #' @field source_url Website address for explanation/introduction of the
+    #' data
+    source_url = "https://covid19cubadata.github.io/#cuba", 
+
 
     #' @description Set up a table of region codes for clean data
     #' @importFrom tibble tibble
