+++ man/Canada.Rd
@@ -57,6 +57,11 @@ Subnational data sources
 across levels.}
 
 \item{\code{source_data_cols}}{existing columns within the raw data}
+
+\item{\code{source_text}}{Plain text description of the source of the data}
+
+\item{\code{source_url}}{Website address for explanation/introduction of the
+data}
 }
 \if{html}{\out{</div>}}
 }
