### 2021-06-26 bb115ca
Fixed handling of NULL in summary table, completed country data
