+++ data-raw/render_available_datasets.R
@@ -1,4 +1,4 @@
 # render available datasets table
 
-all_country_data <- get_regional_data(render = TRUE)
+all_country_data <- get_available_datasets(render = TRUE)
 usethis::use_data(all_country_data, overwrite = TRUE)
