+++ R/Germany.R
@@ -36,6 +36,12 @@ Germany <- R6::R6Class("Germany",
     # nolint end
     #' @field source_data_cols existing columns within the raw data
     source_data_cols = c("cases_new", "deaths_new"),
+    #' @field source_text Plain text description of the source of the data
+    source_text = "Robert Koch-Institut (RKI)",
+    #' @field source_url Website address for explanation/introduction of the
+    #' data
+    source_url = "https://hub.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0/explore",
+
 
     #' @description Set up a table of region codes for clean data
     #' @importFrom tibble tibble
