+++ R/Colombia.R
@@ -35,10 +35,10 @@ Colombia <- R6::R6Class("Colombia",
     #' @field source_data_cols existing columns within the raw data
     source_data_cols = c("cases_total"),
     #' @field source_text Plain text description of the source of the data
-    source_text = "Daniel Cárdenas",
+    source_text = "Daniel C\u00e1rdenas",
     #' @field source_url Website address for explanation/introduction of the
     #' data
-    source_url = "https://github.com/danielcs88/colombia_covid-19/", 
+    source_url = "https://github.com/danielcs88/colombia_covid-19/",
 
     #' @description Set up a table of region codes for clean data
     #' @importFrom tibble tibble
