+++ R/Canada.R
@@ -42,7 +42,7 @@ Canada <- R6::R6Class("Canada",
     #' @field source_url Website address for explanation/introduction of the
     #' data
     source_url = "https://open.canada.ca/data/en/dataset/261c32ab-4cfd-4f81-9dea-7b64065690dc", # nolint
-    
+  
     #' @description Set up a table of region codes for clean data
     #' @importFrom tibble tibble
     set_region_codes = function() {
