+++ R/Netherlands.R
@@ -40,7 +40,7 @@ Netherlands <- R6::R6Class("Netherlands",
     #' @field source_data_cols existing columns within the raw data
     source_data_cols = c("cases_new", "deaths_new", "hosp_new"),
     #' @field source_text Plain text description of the source of the data
-    source_text = "National Institute for Public Health and the Environment (RIVM), Netherlands",
+    source_text = "National Institute for Public Health and the Environment (RIVM), Netherlands", # nolint
     #' @field source_url Website address for explanation/introduction of the
     #' data
     source_url = "https://data.rivm.nl/covid-19/",
