### 2021-06-28 de74f9b
made recomended changes, removed old logos and placed hexsticker in correct location
