+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: covidregionaldata
-Title: Subnational Data for the COVID-19 Outbreak
+Title: covidregionaldata: Subnational data for COVID-19 epidemiology
 Version: 0.9.2.5000
 Authors@R:
-    c(person(given = "Joe",
+    c(person(given = "Joseph",
              family = "Palmer",
              role = "aut",
              email = "joe.palmer.2019@live.rhul.ac.uk",
