### 2021-06-29 c60fb7e
updated version to description and news, also made recomended changes
