+++ NEWS.md
@@ -1,3 +1,7 @@
+# covidregionaldata 0.9.3
+
+This release completes all recommendations for the submission to JOSS.
+
 # covidregionaldata 0.9.2
 
 This release is in development and therefore new features may change without warning (though we will seek to keep this to a minimum).
@@ -9,8 +13,6 @@ This release is in development and therefore new features may change without war
 * Support added for data from the Netherlands provided by RVIM (English: National Institute for Public Health and the Environment). This source provides case, deaths and hospital admission data at the province and municipal levels (by @joseph-palmer).
 * Support added for data from Switzerland and Liechtenstein collated by Canton Zurich (@OpenZH). This source provides case, deaths and hospital admission data at the canton level (by @RichardMN).
 * Made package changes recomended in the JOSS review, including additional statements of need to the README, updates to the manuscript (paper.md) and fixes a bug of multiple sources for some countries. We are very grateful for the detailed feedback given by the JOSS reviewers and their help in improving this package.
-* Package citation has been updated with the JOSS paper citation.
-* Hexsticker has also been added.
 
 ## Changes to implemented data sources
 
