### 2021-06-29 40d0b9d
Add source fields for country and global sources.
