+++ R/SouthAfrica.R
@@ -42,7 +42,6 @@ SouthAfrica <- R6::R6Class("SouthAfrica",
 
     #' @description Set up a table of region codes for clean data
     #' @importFrom tibble tibble
-    #' @importFrom dplyr mutate
     set_region_codes = function() {
       self$codes_lookup$`1` <- tibble(
         code = c(
