+++ .github/workflows/check-full.yaml
@@ -291,5 +291,5 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           context: .
-          push: true
+          file: .devcontainer/Dockerfile          push: true
           tags: docker.pkg.github.com/epiforecasts/covidregionaldata/covidregionaldata:latest
