+++ NEWS.md
@@ -1,6 +1,8 @@
-# covidregionaldata 0.9.3
+# covidregionaldata 0.9.2
 
-This release is in development and therefore new features may change without warning (though we will seek to keep this to a minimum).
+This release adds support for the Covid19 Data Hub which includes Google and Apple mobility data amongst a large range of other data sets, data from the European Commission's Joint Research Centre which is at both the regional and national level, and individual sources for regional data from several countries. Package updates have been made in line with a software review at the [Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/3290). Finally, this release exposes more of the testing infrastructure to users and adds a package hexsticker. 
+
+Thanks to @joseph-palmer, @RichardMN, and @kathsherratt for contributions towards this release.
 
 ## New features
 
@@ -19,6 +21,7 @@ This release is in development and therefore new features may change without war
 * Testing of classes updated to allow for at least one of `common_data_urls` or `level_data_urls` to be present. The previous default which forced the presence of `common_data_urls` meant that several classes had to define an empty field (by @joseph-palmer).
 * Tests on data sets are now included as a method in `DataClass`. `test_regional-datasets` now calls the test function for all classes at each level. Data set specific tests (such as for NHS regions in the UK) are included as a `specific_tests` function within the country class, which is called by the parent (DataClass) `test` after performing standard checks. This allows all the code about a country to be defined in its own class. In addition, users can run tests interactively by calling the test method (e.g. `$test()`;) (by @joseph-palmer)
 * A function to create a template class and automatically add a github workflow file has been added. This makes adding a new data source for a country even easier as now you can call the function `make_new_data_source()` with the country / source name to add and it will set up the basic structure for you. There is also now a github check to make sure all new sources have a workflow set up (by @joseph-palmer).
+* Adds `source_` fields to all data sets to help users properly attribute their data sources (by @RichardMN).
 
 ## Bug fixes
 
