+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: covidregionaldata
 Title: Subnational Data for COVID-19 Epidemiology
-Version: 0.9.3
+Version: 0.9.2
 Authors@R:
     c(person(given = "Joseph",
              family = "Palmer",
