+++ .devcontainer/Dockerfile
@@ -34,10 +34,10 @@ RUN install2.r --error --skipinstalled --repos ${CRAN} --ncpus -1 \
         git2r \
     && rm -rf /tmp/downloaded_packages
 
-# [Optional] Uncomment this section to install additional OS packages.
-# RUN apt-get update \
-#     && export DEBIAN_FRONTEND=noninteractive \
-#     && apt-get -y install --no-install-recommends <your-package-list-here>
+
+RUN apt-get update \
+     && export DEBIAN_FRONTEND=noninteractive \
+     && apt-get -y install --no-install-recommends libgdal-dev
 
 # install dependencies
 COPY DESCRIPTION /tmp/package/DESCRIPTION
