### 2021-07-01 0e7e837
tweak spelling and lint pass
