+++ cran-comments.md
@@ -1,8 +1,0 @@
-## Test environments
-* local R installation, R 4.0.3
-* ubuntu 16.04 (on GitHub actions), R 4.0.3
-* win-builder (devel)
-
-## R CMD check results
-
-0 errors | 0 warnings | 0 notes
