+++ README.md
@@ -1,8 +1,6 @@
 
 # Subnational data for the COVID-19 outbreak <img src="man/figures/logo.png" align="right" alt="" height="150" />
 
-[![Lifecycle:
-maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html)
 [![R-CMD-check](https://github.com/epiforecasts/covidregionaldata/workflows/R-CMD-check/badge.svg)](https://github.com/epiforecasts/covidregionaldata/actions)
 [![Codecov test
 coverage](https://codecov.io/gh/epiforecasts/covidregionaldata/branch/master/graph/badge.svg)](https://codecov.io/gh/epiforecasts/covidregionaldata?branch=master)
@@ -18,8 +16,9 @@ contributors](https://img.shields.io/github/contributors/epiforecasts/covidregio
 [![PRs
 Welcome](https://img.shields.io/badge/PRs-welcome-yellow.svg)](https://makeapullrequest.com/)
 [![GitHub
-commits](https://img.shields.io/github/commits-since/epiforecasts/covidregionaldata/v0.9.1.svg?color=orange)](https://GitHub.com/epiforecasts/covidregionaldata/commit/master/)
+commits](https://img.shields.io/github/commits-since/epiforecasts/covidregionaldata/v0.9.2.svg?color=orange)](https://GitHub.com/epiforecasts/covidregionaldata/commit/master/)
 [![DOI](https://zenodo.org/badge/271601189.svg)](https://zenodo.org/badge/latestdoi/271601189)
+
 [![status](https://joss.theoj.org/papers/dd6f7acdae3b7136a3ac373ce9a0655c/status.svg)](https://joss.theoj.org/papers/dd6f7acdae3b7136a3ac373ce9a0655c)
 
 Interface to subnational and national level COVID-19 data sourced from
@@ -47,7 +46,9 @@ install.packages("covidregionaldata")
 Install the stable development version of the package with:
 
 ``` r
-install.packages("covidregionaldata", repos = "https://epiforecasts.r-universe.dev")
+install.packages("covidregionaldata",
+  repos = "https://epiforecasts.r-universe.dev"
+)
 ```
 
 Install the unstable development version of the package with:
@@ -78,7 +79,7 @@ the temporary directory by default),
 
 ``` r
 start_using_memoise()
-#> Using a cache at: /tmp/RtmpXdh1Cd
+#> Using a cache at: /tmp/RtmpB2UKDX
 ```
 
 To stop using `memoise` use,
@@ -103,19 +104,19 @@ the Google COVID-19 open data project), use:
 ``` r
 nots <- get_national_data()
 #> Downloading data from https://covid19.who.int/WHO-COVID-19-global-data.csv
-#> Rows: 128,546
-#> Columns: 8
+#> Rows: 129165 Columns: 8
+#> ── Column specification ────────────────────────────────────────────────────────
 #> Delimiter: ","
-#> chr  [3]: Country_code, Country, WHO_region
-#> dbl  [4]: New_cases, Cumulative_cases, New_deaths, Cumulative_deaths
-#> date [1]: Date_reported
+#> chr  (3): Country_code, Country, WHO_region
+#> dbl  (4): New_cases, Cumulative_cases, New_deaths, Cumulative_deaths
+#> date (1): Date_reported
 #> 
-#> Use `spec()` to retrieve the guessed column specification
-#> Pass a specification to the `col_types` argument to quiet this message
+#> ℹ Use `spec()` to retrieve the full column specification for this data.
+#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
 #> Cleaning data
 #> Processing data
[…]
