+++ .pre-commit-config.yaml
@@ -15,6 +15,8 @@ repos:
         exclude: >
           (?x)^(
           data/.*| 
+          \.devcontainer/.* |
+          \.github/.*|
           (.*/|)\.Rprofile|
           (.*/|)\.Renviron|
           (.*/|)\.gitignore|
