+++ inst/paper/paper.bib
@@ -33,15 +33,15 @@ urldate = {2021-05-11}
 }
 
 @Article{covidregionaldata,
-  title = {covidregionaldata: Subnational Data for the COVID-19 Outbreak},
-  author = {Sam Abbott and Katharine Sherratt and Joe Palmer and Richard Martin-Nielsen and Jonnie Bevan and Hamish Gibbs and Sebastian Funk},
-  journal = {-},
-  year = {2020},
+  title = {covidregionaldata: Subnational data for COVID-19 epidemiology},
+  author = {Joseph Palmer and Katharine Sherratt and Richard Martin-Nielsen and Jonnie Bevan and Hamish Gibbs and Sebastian Funk and Sam Abbott},
+  journal = {Journal of Open Source Software},
+  year = {2021},
   volume = {-},
   number = {-},
   pages = {-},
   doi = {10.5281/zenodo.3957539},
-}
+  }
 
 @Article{covid19datahub:2020,
   title = {COVID-19 Data Hub},
