+++ R/UK.R
@@ -385,38 +385,43 @@ UK <- R6::R6Class("UK",
       # 1. Data from 7 April 2021 to now:
       # Data not always daily; set up to try urls for last 7 days
       try_date_seq <- seq.Date(self$release_date,
-                               by = -1, length.out = 7)
-      try_urls <- map_chr(try_date_seq,
-                          ~ paste0(
-                              self$data_urls[["nhs_recent_url"]],
-                              "/wp-content/uploads/sites/2/",
-                              year(.x), "/",
-                              ifelse(month(.x) < 10,
-                                     paste0(0, month(.x)),
-                                     month(.x)
-                              ),
-                              "/COVID-19-daily-admissions-and-beds-",
-                              gsub("-", "", as.character(.x)),
-                              ".xlsx")
-                          )
+        by = -1, length.out = 7
+      )
+      try_urls <- map_chr(
+        try_date_seq,
+        ~ paste0(
+          self$data_urls[["nhs_recent_url"]],
+          "/wp-content/uploads/sites/2/",
+          year(.x), "/",
+          ifelse(month(.x) < 10,
+            paste0(0, month(.x)),
+            month(.x)
+          ),
+          "/COVID-19-daily-admissions-and-beds-",
+          gsub("-", "", as.character(.x)),
+          ".xlsx"
+        )
+      )
 
       names(try_urls) <- try_date_seq
       # Check for working urls
-      url_status <- map_chr(try_urls,
-                        ~ GET(.x) %>%
-                          status_code())
+      url_status <- map_chr(
+        try_urls,
+        ~ GET(.x) %>%
+          status_code()
+      )
       # Keep latest working url
       url_status <- url_status[(url_status == 200)]
       names(url_status) <- as.Date(names(url_status))
       nhs_recent_url <- try_urls[as.character(max(names(url_status)))]
       # Get latest url download
       recent <- download_excel(nhs_recent_url,
-                               "nhs_recent.xlsx",
-                               verbose = self$verbose,
-                               transpose = TRUE,
-                               sheet = 1,
-                               range = cell_limits(c(28, 2), c(36, NA))
-        )
+        "nhs_recent.xlsx",
+        verbose = self$verbose,
+        transpose = TRUE,
+        sheet = 1,
+        range = cell_limits(c(28, 2), c(36, NA))
+      )
 
       # 2. Data for August 2020 to 7 April 2021
       archive <- download_excel(
