### 2021-07-01 #3290 @seabbs
Thanks @csoneson (standing in briefly for @joseph-palmer). I think we have dealt with the outstanding issues 😄.

Version tag: `0.9.2`
Tagged release: https://github.com/epiforecasts/covidregionaldata/releases/tag/0.9.2
Archived DOI: 10.5281/zenodo.5055487
- [x] Check metadata
- [x] Fixed Zenodo reference in paper.
