### 2021-07-02 #3290 @csoneson
Thanks @seabbs. The Zenodo archive looks good! However, I'm not sure about the citation to Palmer et al (previously Abbott et al) - the paper is basically citing itself, but with a URL pointing to Zenodo. I would suggest removing the citation (or make it a website citation, not pointing to the JOSS publication) - the link to Zenodo is still there (note though that currently the link does not point to the URL given in the citation), and the version above is also explicitly added also to the published record (there will be a link to a 'Software Archive' on the paper's landing page).
