### 2021-07-02 b7eeeaa
Remove zenodo links and reference
