### 2021-07-02 #3290 @seabbs
Thanks @csoneson

I have removed the citation and link to Zenodo in the draft, deleted and recreated the release, and re-edited the zenodo release to share meta data with JOSS.

Zenodo: https://zenodo.org/record/5059988
Doi: 10.5281/zenodo.5059988
