### 2021-07-03 #3290 @arfon
@joseph-palmer – please merge this change which fixes the footnotes for your paper.
