### 2021-07-05 #3290 @Kevin-Mattheus-Moerman
Congratulation @joseph-palmer this work is now published in JOSS!

Thanks @mponce0 and @federicomarini for reviewing this work!!!

Also thank you @csoneson for editing.
